% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_to_list}
\alias{args_to_list}
\title{Get all args of parent call (both specified and defaults) as list}
\usage{
args_to_list(call = sys.call(sys.parent()), fn = sys.function(sys.parent()))
}
\arguments{
\item{call}{do not use}

\item{fn}{do not use}
}
\value{
a list of all for the parent function call
}
\description{
Get all args of parent call (both specified and defaults) as list
}
