% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Truncates predictions to ensure loss function is bounded.}
\usage{
bound(x, bounds = 0.001)
}
\arguments{
\item{bounds}{Either a length two vector of c(lower,upper) or a lower bound,
where the upper is then 1 - lower. Default bounds=0.001.}
}
\value{
truncated predictions.
}
\description{
Truncates predictions to ensure loss function is bounded.
}
