% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sl3-package.R
\docType{data}
\name{bsds}
\alias{bsds}
\title{Bicycle sharing time series dataset}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset}

Fanaee-T, Hadi, and Gama, Joao, 'Event labeling combining ensemble detectors and background knowledge', Progress in Artificial Intelligence (2013): pp. 1-15, Springer Berlin Heidelberg
}
\usage{
bsds
}
\description{
Bicycle sharing time series dataset from the UCI Machine Learning Repository.
}
\examples{
data(bsds)
head(bsds)
}
\keyword{data}

