% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3-package.R
\docType{data}
\name{cpp_1yr}
\alias{cpp_1yr}
\title{Subset of growth data from the collaborative perinatal project (CPP)}
\source{
\url{https://catalog.archives.gov/id/606622}

Broman, Sarah. 'The collaborative perinatal project: an overview.' Handbook of longitudinal research 1 (1984): 185-227.
}
\usage{
data(cpp_1yr)
}
\description{
Subset of growth data from the collaborative perinatal project (CPP) at single time-point. 
The rows in original \code{cpp} data were subset for agedays==366.
See \code{?cpp} for the description of the variables.
}
\examples{
data(cpp_1yr)
head(cpp_1yr)
table(cpp_1yr[["agedays"]])
}
\keyword{data}
