% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_chain.R
\name{customize_chain}
\alias{customize_chain}
\title{Modify the chain behavior of a learner}
\usage{
customize_chain(learner, chain_fun)
}
\arguments{
\item{learner}{sl3 learner to modify}

\item{chain_fun}{function with arguments learner and task that defines the new chain behavior}
}
\description{
This function wraps a learner in such a way that the behavior of learner$chain is modified to use a new function definition. learner$train and learner$predict are unaffected
}
