% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_cv.R
\name{train_task}
\alias{train_task}
\alias{validation_task}
\title{Subset Tasks for CV
THe functions use origami folds to subset tasks. These functions are used by Lrnr_cv
(and therefore other learners that use Lrnr_cv). So that nested cv works properly, currently
the subsetted task objects do not have fold structures of their own, and so generate them from
defaults if nested cv is requested.}
\usage{
train_task(task, fold)

validation_task(task, fold)
}
\arguments{
\item{task}{a task to subset}

\item{fold}{an origami fold object to use for subsetting}
}
\description{
Subset Tasks for CV
THe functions use origami folds to subset tasks. These functions are used by Lrnr_cv
(and therefore other learners that use Lrnr_cv). So that nested cv works properly, currently
the subsetted task objects do not have fold structures of their own, and so generate them from
defaults if nested cv is requested.
}
