% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Learner.R
\docType{class}
\name{CV_Learner}
\alias{CV_Learner}
\title{Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions}
\format{\code{\link{R6Class}} object.}
\usage{
CV_Learner
}
\value{
\code{\link{Learner}} object with methods for training and prediction
}
\description{
Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions
}
\section{Fields}{

\describe{
\item{\code{params}}{A learner to be wrapped}
}}

\section{Methods}{

\describe{
  \item{\code{new(learner)}}{Wraps the \code{learner} in a cross-validation layer.}
  }
}

\seealso{
Other Learners: \code{\link{GLM_Learner}},
  \code{\link{Learner}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
