% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{cv_risk}
\alias{cv_risk}
\title{Cross-validated Risk Estimation}
\usage{
cv_risk(learner, loss_fun, coefs = NULL)
}
\arguments{
\item{learner}{A trained learner object.}

\item{loss_fun}{A valid loss function. See \code{\link{loss_functions}}.}

\item{coefs}{A \code{numeric} vector of coefficients.}
}
\description{
Estimates the cross-validated risk for a given learner and loss function.
}
