% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_metalearner.R
\name{default_metalearner}
\alias{default_metalearner}
\title{Automatically Defined Metalearner}
\usage{
default_metalearner(outcome_type)
}
\arguments{
\item{outcome_type}{a Variable_Type object}
}
\description{
A sensible metalearner based on \code{\link{Lrnr_solnp}} is chosen based on
outcome type. This amounts to choosing an appropriate loss function and
combination function.
}
\details{
\tabular{lll}{
   Outcome Type \tab Combination Function \tab Loss Function \cr
   binomial \tab metalearner_logistic_binomial \tab loss_squared_error \cr
   categorical \tab metalearner_linear_multinomial \tab loss_loglik_multinomial \cr
   continuous \tab metalearner_linear \tab loss_squared_error \cr
   multivariate \tab metalearner_linear_multivariate \tab loss_squared_error_multivariate \cr
}
}
