% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3-package.R
\docType{data}
\name{density_dat}
\alias{density_dat}
\title{Simulated data with continuous exposure}
\usage{
data(density_dat)
}
\description{
Simulated data with continuous exposure, used with examples of conditional
density estimation.
}
\examples{
data(density_dat)
head(density_dat)
#
}
\keyword{data}
