% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_sl.R
\name{drop_offsets_chain}
\alias{drop_offsets_chain}
\title{Chain while dropping offsets}
\usage{
drop_offsets_chain(learner, task)
}
\arguments{
\item{task}{An object of class \code{sl3_Task}.}
}
\description{
Allows the dropping of offsets when calling the chain method. This is simply
a modified version of the chain method found in \code{Lrnr_base}. INTERNAL
USE ONLY.
}
\keyword{internal}
