% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_utils.R
\name{estQuantile}
\alias{estQuantile}
\title{Generate quantiles estimates from the estimated CDF}
\usage{
estQuantile(task, hazard_fit, qProb = list(0.05, 0.95))
}
\arguments{
\item{task}{A \code{\link{sl3_Task}} where the outcome is cdf estimate.}

\item{hazard_fit}{A \code{\link{sl3}} fit for the cdf.}

\item{qProb}{A list of quantiles}
}
\description{
Generate quantiles estimates from the estimated CDF
}
