% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_helpers.R
\name{factor_to_indicators}
\alias{factor_to_indicators}
\alias{dt_expand_factors}
\title{Convert Factors to indicators}
\usage{
factor_to_indicators(x, ind_ref_mat = NULL)

dt_expand_factors(dt)
}
\arguments{
\item{x}{the factor to expand}

\item{ind_ref_mat}{a matrix used for expansion, if NULL generated automatically}

\item{dt}{the dt to expand}
}
\description{
replicates the functionality of model.matrix, but faster

Replicates the functionality of \code{model.matrix}, but faster
}
