% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_glm_family}
\alias{get_glm_family}
\title{determine glm family}
\usage{
get_glm_family(family, outcome_type)
}
\arguments{
\item{family}{family argument specified to learner, either character or an actual family object}

\item{outcome_type}{outcome type from Learner$get_outcome_type}
}
\value{
family object determined by combination of arguments
}
\description{
determine glm family
}
