% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_randomForest.R
\name{importance}
\alias{importance}
\title{Importance
Extract variable importance measures produced by
\code{\link[randomForest]{randomForest}} and order in decreasing order of
importance.}
\usage{
importance(type = ...)
}
\arguments{
\item{type}{either 1 or 2, specifying the type of importance measure (1=mean
decrease in accuracy, 2=mean decrease in node impurity).}
}
\description{
Importance
Extract variable importance measures produced by
\code{\link[randomForest]{randomForest}} and order in decreasing order of
importance.
}
