% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{impute}
\alias{impute}
\title{Impute missing values with the median/mode
based on imputeMissings R package (removed from CRAN)}
\usage{
impute(data)
}
\arguments{
\item{data}{A data frame with dummies or numeric variables.}
}
\description{
Character vectors and factors are imputed with the mode.
Numeric and integer vectors are imputed with median.
}
\keyword{internal}
