% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{inverse_sample}
\alias{inverse_sample}
\title{Inverse CDF Sampling}
\usage{
inverse_sample(n_samples, cdf = NULL, pdf = NULL)
}
\arguments{
\item{n_samples}{If \code{true}, remove entries after failure time for each
observation.}

\item{cdf}{A list with x and y representing the cdf}

\item{pdf}{A list with x and y representing the pdf}
}
\description{
Inverse CDF Sampling
}
