% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_only_fun_args}
\alias{keep_only_fun_args}
\title{Streamline Function Arguments}
\usage{
keep_only_fun_args(Args, fun)
}
\arguments{
\item{Args}{A \code{list} of function arguments to be streamlined.}

\item{fun}{A \code{function} whose signature will be used to reduce the
arguments passed in.}
}
\description{
Reduce a list of function argsuments by taking a function body and returning
only the arguments that belong to the function signature.
}
\keyword{internal}
