% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_learners.R
\name{sl3_list_properties}
\alias{sl3_list_learners}
\alias{sl3_list_properties}
\title{List sl3 Learners}
\usage{
sl3_list_properties()

sl3_list_learners(properties = c())
}
\arguments{
\item{properties}{a vector of properties that learners must match to be returned}
}
\value{
a vector of learner names that match the property list
}
\description{
Lists learners in \code{sl3} (defined as objects that start with \code{Lrnr_} and inherit from \code{Lrnr_base})
}

