% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_arima.R
\docType{class}
\name{Lrnr_arima}
\alias{Lrnr_arima}
\title{Univariate ARIMA Models}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_arima
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner supports autoregressive integrated moving average model for univariate time-series.
}
\section{Fields}{

\describe{
\item{\code{order}}{A specification of the non-seasonal part of the ARIMA model: the three integer components (p, d, q) are the AR order, the degree of differencing, and the MA order.}

\item{\code{seasonal}}{A specification of the seasonal part of the ARIMA model, plus the period (which defaults to frequency(x)). This should be a list with components order and period, but a specification of just a numeric vector of length 3 will be turned into a suitable list with the specification as the order.}

\item{\code{n.ahead}}{The forecast horizon. If not specified, returns forecast of size \code{task$X}.}
}}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_base}}; \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_glm_fast}},
  \code{\link{Lrnr_h2o_classifier}},
  \code{\link{Lrnr_h2o_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_h2o_mutator}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_pkg_SuperLearner_screener}},
  \code{\link{Lrnr_randomForest}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_xgboost}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_cv}}; \code{\link{Lrnr_delayed}};
  \code{\link{Lrnr_expSmooth}}; \code{\link{Lrnr_rugarch}};
  \code{\link{Lrnr_tsDyn}}; \code{\link{Pipeline}};
  \code{\link{Stack}}
}
\keyword{data}

