% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_condensier.R
\docType{class}
\name{Lrnr_condensier}
\alias{Lrnr_condensier}
\title{Conditional Density Estimation}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_condensier
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction.
}
\description{
This learner provides facilities for conditional density estimation using
the \code{condensier} package.
}
\section{Fields}{

\describe{
\item{\code{bin_method}}{The type of smoothing to be performed. See documentation of
the \code{condensier} package for details.}

\item{\code{nbins}}{The number of observations per bin. See documentation of the
\code{condensier} package for details.}

\item{\code{max_n_cat}}{Maximum number of unique levels for categorical outcomes.
See documentation of the \code{condensier} package for details.}

\item{\code{pool}}{Whether pooling of data across bins should be performed. See
documentation of the \code{condensier} package for details.}

\item{\code{max_n_bin}}{Maximum number of observations per single bin for continuous
outcome. See documentation of the \code{condensier} package for details.}

\item{\code{parfit}}{Whether to invoke parallelization in the fitting procedure. See
documentation of the \code{condensier} package for details.}

\item{\code{bin_estimator}}{The classification algorithm to be used in the fitting
process. See documentation of the \code{condensier} package for details.}

\item{\code{intrvls}}{An interval range to be used for custom bin definitions. See
 documentation of the \code{condensier} package for details.}
}}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_arima}}; \code{\link{Lrnr_base}};
  \code{\link{Lrnr_cv}};
  \code{\link{Lrnr_define_interactions}};
  \code{\link{Lrnr_expSmooth}}; \code{\link{Lrnr_rugarch}};
  \code{\link{Lrnr_sl}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_tsDyn}}; \code{\link{Pipeline}};
  \code{\link{Stack}}
}
\keyword{data}

