% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_cv.R
\docType{class}
\name{Lrnr_cv}
\alias{Lrnr_cv}
\title{Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_cv
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions
}
\section{Fields}{

\describe{
\item{\code{params}}{A learner to be wrapped}
}}
\section{Methods}{

\describe{
  \item{\code{new(learner)}}{Wraps the \code{learner} in a cross-validation layer.}
  }
}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_arima}}; \code{\link{Lrnr_base}};
  \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_glm_fast}},
  \code{\link{Lrnr_h2o_classifier}},
  \code{\link{Lrnr_h2o_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_h2o_mutator}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_pkg_SuperLearner_screener}},
  \code{\link{Lrnr_randomForest}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_xgboost}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_delayed}}; \code{\link{Lrnr_rugarch}};
  \code{\link{Lrnr_tsDyn}}; \code{\link{Pipeline}};
  \code{\link{Stack}}
}
\keyword{data}

