% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_dBart.R
\docType{class}
\name{Lrnr_dBart}
\alias{Lrnr_dBart}
\title{Discrete Bayesian Additive Regression Tree sampler}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_dBart
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner implements BART algorithm in C++, using the \code{dbarts} package.
}
\section{Parameters}{

\describe{
\item{\code{Y}}{Outcome variable}
\item{\code{X}}{Covariate dataframe}
\item{\code{test}}{An optional matrix or data frame with the same number of
predictors as data, or formula in backwards compatibility mode. If column names
are present, a matching algorithm is used.}
\item{\code{subset}}{An optional vector specifying a subset of observations to
be used in the fitting process.}
\item{\code{weights}}{An optional vector of weights to be used in the fitting
process. When present, BART fits a model with observations y | x ~ N(f(x), σ^2 / w),
where f(x) is the unknown function.}
\item{\code{offset}}{An optional vector specifying an offset from 0 for the
relationship between the underyling function, f(x), and the response y.
Only is useful for binary responses, in which case the model fit is to assume
P(Y = 1 | X = x) = Φ(f(x) + offset), where Φ is the standard normal cumulative
distribution function.}
\item{\code{offset.test}}{The equivalent of offset for test observations.
Will attempt to use offset when applicable.}
\item{\code{verbose}}{A logical determining if additional output is printed to
the console. See dbartsControl.}
\item{\code{n.samples}}{A positive integer setting the default number of
posterior samples to be returned for each run of the sampler.
Can be overriden at run-time. See dbartsControl.}
\item{\code{tree.prior}}{An expression of the form cgm or cgm(power, base)
setting the tree prior used in fitting.}
\item{\code{node.prior}}{An expression of the form normal or normal(k) that
sets the prior used on the averages within nodes.}
\item{\code{resid.prior}}{An expression of the form chisq or chisq(df, quant)
that sets the prior used on the residual/error variance.}
\item{\code{control}}{An object inheriting from dbartsControl, created by the dbartsControl function.}
\item{\code{sigma}}{A positive numeric estimate of the residual standard deviation.
If NA, a linear model is used with all of the predictors to obtain one.}
}
}

\section{Common Parameters}{


Individual learners have their own sets of parameters. Below is a list of shared parameters, implemented by \code{Lrnr_base}, and shared
by all learners.

\describe{
  \item{\code{covariates}}{A character vector of covariates. The learner will use this to subset the covariates for any specified task}
  \item{\code{outcome_type}}{A \code{\link{variable_type}} object used to control the outcome_type used by the learner. Overrides the task outcome_type if specified}
  \item{\code{...}}{All other parameters should be handled by the invidual learner classes. See the documentation for the learner class you're instantiating}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_bartMachine}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_bilstm}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pca}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}}, \code{\link{Lrnr_rpart}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Lrnr_xgboost}}, \code{\link{Pipeline}},
  \code{\link{Stack}}, \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
