% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_define_interactions.R
\docType{class}
\name{Lrnr_define_interactions}
\alias{Lrnr_define_interactions}
\title{Conversion to Design Matrices}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_define_interactions
}
\description{
Utility for converting \code{data.frame} with raw data columns into a model
matrix with columns like interaction terms and factor indicators.
TODO: reimplement this without using \code{model.matrix}.
}
\section{Fields}{

\describe{
\item{\code{interactions}}{A character or model formula object identifying the exact
interaction terms to be computed in building a design matrix.}
}}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\concept{Learners}
\keyword{datasets}
