% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_earth.R
\docType{class}
\name{Lrnr_earth}
\alias{Lrnr_earth}
\title{Earth: Multivariate Adaptive Regression Splines}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
\code{\link{Lrnr_base}}.
}
\value{
A learner object inheriting from \code{\link{Lrnr_base}} with
methods for training and prediction. For a full list of learner
functionality, see the complete documentation of \code{\link{Lrnr_base}}.
}
\description{
This learner provides fitting procedures for building regression models thru
the spline regression techniques described in
\insertCite{friedman1991multivariate;textual}{sl3} and
\insertCite{friedman1993fast;textual}{sl3}, via \pkg{earth} and the function
\code{\link[earth]{earth}}.
}
\section{Parameters}{

\itemize{
\item \code{degree}: A \code{numeric} specifying the maximum degree of
interactions to be used in the model. This defaults to 2, specifying
up through one-way interaction terms. Note that this differs from the
default of \code{\link[earth]{earth}}.
\item \code{penalty}: Generalized Cross Validation (GCV) penalty per knot.
Defaults to 3 as per the recommendation for \code{degree} > 1 in the
documentation of \code{\link[earth]{earth}}. Special values (for use
by knowledgeable users): The value 0 penalizes only terms, not knots.
The value -1 translates to no penalty.
\item \code{pmethod}: Pruning method, defaulting to \code{"backward"}. Other
options include \code{"none"}, \code{"exhaustive"}, \code{"forward"},
\code{"seqrep"}, \code{"cv"}.
\item \code{nfold}: Number of cross-validation folds. The default is 0, for no
cross-validation.
\item \code{ncross}: Only applies if \code{nfold} > 1, indicating the number
of cross-validation rounds. Each cross-validation has \code{nfold}
folds. Defaults to 1.
\item \code{minspan}: Minimum number of observations between knots.
\item \code{endspan}: Minimum number of observations before the first and
after the final knot.
\item \code{...}: Other parameters passed to \code{\link[earth]{earth}}. See
its documentation for details.
}
}

\examples{
data(cpp_imputed)
covars <- c(
  "apgar1", "apgar5", "parity", "gagebrth", "mage", "meducyrs", "sexn"
)
outcome <- "haz"
task <- sl3_Task$new(cpp_imputed,
  covariates = covars,
  outcome = outcome
)
# fit and predict from a MARS model
earth_lrnr <- make_learner(Lrnr_earth)
earth_fit <- earth_lrnr$train(task)
earth_preds <- earth_fit$predict(task)
}
\references{
\insertAllCited{}
}
\seealso{
Other Learners: 
\code{\link{Custom_chain}},
\code{\link{Lrnr_HarmonicReg}},
\code{\link{Lrnr_arima}},
\code{\link{Lrnr_bartMachine}},
\code{\link{Lrnr_base}},
\code{\link{Lrnr_bayesglm}},
\code{\link{Lrnr_bilstm}},
\code{\link{Lrnr_bound}},
\code{\link{Lrnr_caret}},
\code{\link{Lrnr_cv_selector}},
\code{\link{Lrnr_cv}},
\code{\link{Lrnr_dbarts}},
\code{\link{Lrnr_define_interactions}},
\code{\link{Lrnr_density_discretize}},
\code{\link{Lrnr_density_hse}},
\code{\link{Lrnr_density_semiparametric}},
\code{\link{Lrnr_expSmooth}},
\code{\link{Lrnr_gam}},
\code{\link{Lrnr_gbm}},
\code{\link{Lrnr_glm_fast}},
\code{\link{Lrnr_glmnet}},
\code{\link{Lrnr_glm}},
\code{\link{Lrnr_grf}},
\code{\link{Lrnr_gru_keras}},
\code{\link{Lrnr_gts}},
\code{\link{Lrnr_h2o_grid}},
\code{\link{Lrnr_hal9001}},
\code{\link{Lrnr_haldensify}},
\code{\link{Lrnr_hts}},
\code{\link{Lrnr_independent_binomial}},
\code{\link{Lrnr_lightgbm}},
\code{\link{Lrnr_lstm_keras}},
\code{\link{Lrnr_mean}},
\code{\link{Lrnr_multiple_ts}},
\code{\link{Lrnr_multivariate}},
\code{\link{Lrnr_nnet}},
\code{\link{Lrnr_nnls}},
\code{\link{Lrnr_optim}},
\code{\link{Lrnr_pca}},
\code{\link{Lrnr_pkg_SuperLearner}},
\code{\link{Lrnr_polspline}},
\code{\link{Lrnr_pooled_hazards}},
\code{\link{Lrnr_randomForest}},
\code{\link{Lrnr_ranger}},
\code{\link{Lrnr_revere_task}},
\code{\link{Lrnr_rpart}},
\code{\link{Lrnr_rugarch}},
\code{\link{Lrnr_screener_augment}},
\code{\link{Lrnr_screener_coefs}},
\code{\link{Lrnr_screener_correlation}},
\code{\link{Lrnr_screener_importance}},
\code{\link{Lrnr_sl}},
\code{\link{Lrnr_solnp_density}},
\code{\link{Lrnr_solnp}},
\code{\link{Lrnr_stratified}},
\code{\link{Lrnr_subset_covariates}},
\code{\link{Lrnr_svm}},
\code{\link{Lrnr_tsDyn}},
\code{\link{Lrnr_ts_weights}},
\code{\link{Lrnr_xgboost}},
\code{\link{Pipeline}},
\code{\link{Stack}},
\code{\link{define_h2o_X}()},
\code{\link{undocumented_learner}}
}
\concept{Learners}
