% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_glm.R
\docType{class}
\name{Lrnr_glm}
\alias{Lrnr_glm}
\title{GLM Fits}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_glm
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides fitting procedures for generalized linear models by way
of a wrapper relying on \code{stats::glm}.
}
\section{Fields}{

\describe{
\item{\code{family}}{A \code{family} object from package \code{stats} describing the
error family of the model to be fit. See the documentation for the package
\code{stats} for details, or consult \code{stats::family} directly.}

\item{\code{...}}{Additional arguments.}
}}

\keyword{data}
