% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_glm_fast.R
\docType{class}
\name{Lrnr_glm_fast}
\alias{Lrnr_glm_fast}
\title{Faster GLM Fits}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_glm_fast
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides faster fitting procedures for generalized linear models
by way of a wrapper relying on the \code{speedglm} package.
}
\section{Fields}{

\describe{
\item{\code{family}}{A \code{family} object from package \code{stats} describing the
error family of the model to be fit. See the documentation for the package
\code{speedglm} for details.}

\item{\code{method}}{The type of matrix decomposition to be used in the model
fitting process. See documentation for the package \code{speedglm} for
further details.}

\item{\code{covariates}}{Extra covariate terms to be passed to the model fitting
process. See documentation of the \code{speedglm} package for details.}

\item{\code{...}}{Additional arguments.}
}}
\keyword{data}

