% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_HarmonicReg.R
\docType{class}
\name{Lrnr_HarmonicReg}
\alias{Lrnr_HarmonicReg}
\title{Harmonic Regression}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_HarmonicReg
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner fits first harmonics in a Fourier expansion to one or more time series.
}
\section{Fields}{

\describe{
\item{\code{order}}{A specification of the non-seasonal part of the ARIMA model: the three integer components (p, d, q) are the AR order, the degree of differencing, and the MA order.}

\item{\code{seasonal}}{A specification of the seasonal part of the ARIMA model, plus the period (which defaults to frequency(x)). This should be a list with components order and period, but a specification of just a numeric vector of length 3 will be turned into a suitable list with the specification as the order.}

\item{\code{n.ahead}}{The forecast horizon. If not specified, returns forecast of size task$X.}

\item{\code{K}}{Maximum order(s) of Fourier terms.}

\item{\code{freq}}{the number of observations per unit of time.}
}}

\seealso{
Other Learners: \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
