% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_independent_binomial.R
\docType{class}
\name{Lrnr_independent_binomial}
\alias{Lrnr_independent_binomial}
\title{Classification from Binomial Regression}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_independent_binomial
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides converts a binomial learner into a multinomial learner using a series of independent binomials. 
The procedure is modelled on \url{https://en.wikipedia.org/wiki/Multinomial_logistic_regression#As_a_set_of_independent_binary_regressions}
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
