% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_mean.R
\docType{class}
\name{Lrnr_mean}
\alias{Lrnr_mean}
\title{Intercept Model Fits}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_mean
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction.
}
\description{
This learner provides fitting procedures for intercept models. Such models
predict the outcome variable simply as the mean of the outcome vector.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments.}
}}

\keyword{data}
