% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_nnls.R
\docType{class}
\name{Lrnr_nnls}
\alias{Lrnr_nnls}
\title{Non-negative Linear Least Squares}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_nnls
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction.
}
\description{
This learner provides fitting procedures for models with non-negative linear
least squares, internally using the \code{nnls} package. For details on the
fitting procedure, consult the documentation of the \code{nnls} package.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
