% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_optim.R
\docType{class}
\name{Lrnr_optim}
\alias{Lrnr_optim}
\title{Optimization}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_optim
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This meta-learner provides fitting procedures for any pairing of loss 
function and metalearner function, subject to constraints. The
optimization problem is solved by making use of \code{optim}, For further details, 
consult the documentation of \code{\link{optim}}
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
