% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pca.R
\docType{class}
\name{Lrnr_pca}
\alias{Lrnr_pca}
\title{Principal Component Analysis and Regression}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_pca
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides facilities for performing principal components analysis
(PCA) to reduce the dimensionality of a data set to a pre-specified value.
For further details, consult the documentation of \code{prcomp} from the core
package \code{stats}. This learner object is primarily intended for use with
other learners as part of a pre-processing pipeline.
}
\section{Parameters}{

\describe{
\item{\code{n_comp}}{A \code{numeric} value indicating the number of
components to be produced as a result of the PCA dimensionality
reduction. For convenience, this defaults to two (2) components.}
\item{\code{center}}{A \code{logical} value indicating whether the input
data matrix should be centered before performing PCA. This defaults to
\code{TRUE} since that is the recommended practice. Consider consulting
the documentation of \code{prcomp} for details.}
\item{\code{scale.}}{A \code{logical} value indicating whether the input
data matrix should be scaled (to unit variance) before performing PCA.
Consider consulting the documentation of \code{prcomp} for details.}
\item{\code{...}}{Other optional parameters to be passed to \code{prcomp}.
Consider consulting the documentation of \code{prcomp} for details.}
}
}

\section{Common Parameters}{


Individual learners have their own sets of parameters. Below is a list of shared parameters, implemented by \code{Lrnr_base}, and shared
by all learners.

\describe{
  \item{\code{covariates}}{A character vector of covariates. The learner will use this to subset the covariates for any specified task}
  \item{\code{outcome_type}}{A \code{\link{variable_type}} object used to control the outcome_type used by the learner. Overrides the task outcome_type if specified}
  \item{\code{...}}{All other parameters should be handled by the invidual learner classes. See the documentation for the learner class you're instantiating}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_bartMachine}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_bilstm}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}}, \code{\link{Lrnr_rpart}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Lrnr_xgboost}}, \code{\link{Pipeline}},
  \code{\link{Stack}}, \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
