% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_pkg_condensier_logisfitR6.R
\docType{class}
\name{Lrnr_pkg_condensier_logisfitR6}
\alias{Lrnr_pkg_condensier_logisfitR6}
\title{Lrnr_pkg_condensier_logisfitR6}
\format{\preformatted{Class 'R6ClassGenerator' <Lrnr_pkg_condensier_logisfitR6> object generator
  Inherits from: <condensier::logisfitR6>
  Public:
    lmclass: NULL
    fitfunname: NULL
    initialize: function (sl3_lrnr, ...) 
    fit: function (datsum_obj) 
    predict: function (datsum_obj, m.fit) 
    clone: function (deep = FALSE) 
  Private:
    fallback_learner: R6ClassGenerator
    sl3_lrnr: NULL
  Parent env: <environment: namespace:sl3>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "Lrnr_pkg_condensier_logisfitR6_generator"
}}
\usage{
Lrnr_pkg_condensier_logisfitR6
}
\description{
Lrnr_pkg_condensier_logisfitR6
}
\keyword{datasets}

