% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_SuperLearner.R
\docType{class}
\name{Lrnr_pkg_SuperLearner}
\alias{Lrnr_pkg_SuperLearner}
\title{SuperLearner Wrapper Interface}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_pkg_SuperLearner
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides an interface to the wrapper functions provided by the
\code{SuperLearner} package. Such wrapper functions are used to draw on
learning algorithms provided by a wide variety of R packages, towards the
goal of fitting a \code{SuperLearner} ensemble model. Consult documentation
of the \code{SuperLearner} package for details on these wrapper functions.
}
\section{Fields}{

\describe{
\item{\code{SL_wrapper}}{A \code{SuperLearner} wrapper, used as an interface to
algorithms to be fit in a \code{SuperLearner} ensemble model.}

\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
