% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_SuperLearner_method.R
\docType{data}
\name{Lrnr_pkg_SuperLearner_method}
\alias{Lrnr_pkg_SuperLearner_method}
\title{Methods for SuperLearner Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_pkg_SuperLearner_method
}
\description{
Support for \code{SuperLearner} wrappers, obviously not the most efficient
approach, we should reimplement as many as possible.
}
\keyword{datasets}
