% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_SuperLearner_screener.R
\docType{class}
\name{Lrnr_pkg_SuperLearner_screener}
\alias{Lrnr_pkg_SuperLearner_screener}
\title{SuperLearner Screener Interface}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_pkg_SuperLearner_screener
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction.
}
\description{
This learner provides an interface to the screener functions provided by the
\code{SuperLearner} package. Such screening functions are used to reduce the
number of covariates in the process of fitting a \code{SuperLearner} ensemble
model. Consult documentation of the \code{SuperLearner} package for details
on these screener functions.
}
\section{Fields}{

\describe{
\item{\code{SL_wrapper}}{A \code{SuperLearner} wrapper, used as an interface to
algorithms to be fit in a \code{SuperLearner} ensemble model.}

\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
