% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_randomForest.R
\docType{class}
\name{Lrnr_randomForest}
\alias{Lrnr_randomForest}
\title{Random Forest Models}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_randomForest
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction.
}
\description{
This learner provides fitting procedures for random forest models, using the
\code{randomForest} package. For details on the fitting procedure, consult
the documentation of the \code{randomForest} package.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
