% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_rugarch.R
\docType{class}
\name{Lrnr_rugarch}
\alias{Lrnr_rugarch}
\title{Univariate GARCH Models}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_rugarch
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner supports autoregressive fractionally integrated moving average and
various flavors of generalized autoregressive conditional heteroskedasticity models for univariate time-series.
}
\section{Fields}{

\describe{
\item{\code{variance.model}}{List containing the variance model specification. This includes model, GARCH order, submodel, external regressors and variance tageting. Refer to \code{ugarchspec} for more information.}

\item{\code{mean.model}}{List containing the mean model specification. This includes ARMA model, whether the mean should be included, and external regressors among others. Refer to \code{ugarchspec} for more information.}

\item{\code{distribution.model}}{Conditional density to use for the innovations.}

\item{\code{start.pars}}{List of staring parameters for the optimization routine.}

\item{\code{fixed.pars}}{List of parameters which are to be kept fixed during the optimization.}

\item{\code{n.ahead}}{The forecast horizon.}
}}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_arima}}; \code{\link{Lrnr_base}};
  \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_glm_fast}},
  \code{\link{Lrnr_h2o_classifier}},
  \code{\link{Lrnr_h2o_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_h2o_mutator}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_pkg_SuperLearner_screener}},
  \code{\link{Lrnr_randomForest}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_xgboost}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_cv}}; \code{\link{Lrnr_delayed}};
  \code{\link{Lrnr_expSmooth}}; \code{\link{Lrnr_tsDyn}};
  \code{\link{Pipeline}}; \code{\link{Stack}}
}
\keyword{data}

