% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_screener_importance.R
\docType{class}
\name{Lrnr_screener_importance}
\alias{Lrnr_screener_importance}
\title{Random Forest Screener}
\format{
\code{\link{R6Class}} object.
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides covariate screening procedures using variable
importance measures as produced by learners that support variable importance.
}
\section{Parameters}{

\describe{
\item{\code{learner}}{An instantiated learner to use for estimating
variable importance.}
\item{\code{importance_fun}}{Name of the variable importance function.}
\item{\code{model_arg}}{Name of the argument that defines the fitted model.}
\item{\code{importance_fun_control}}{List of additional arguments to be
passed to importance_fun, in addition to the fitted model.}
\item{\code{feature_var}}{Name of importance object column that corresponds
to the covariates}
\item{\code{importance_var}}{Name of importance object column that corresponds
to the importance metric of interest. This variable will be used to sort
and screen variables. If NULL, importance object is assumed to be sorted,
either by default or in user-supplied \code{format_importance_object_fun}
function.}
\item{\code{min_retain = 2}}{Minimum number of variables to be kept.}
\item{\code{max_retain = 10}}{Maximum number of variables to be kept.}
\item{\code{threshold = NULL}}{Minimum size of importance_var to be kept.
If NULL, will select max_retain variables.}
\item{\code{format_importance_object_fun}}{Function to customize
formatting of importance object. This function must return a
\code{data.table} or \code{data.frame}, number of rows must be greater than
max_retain, and \code{feature_var} value as a column name.}
\item{\code{verbose = FALSE}}{Print selected variables.}
\item{\code{...}}{Other parameters passed to \code{learner}.}
}
}

\seealso{
Other Learners: 
\code{\link{Custom_chain}},
\code{\link{Lrnr_HarmonicReg}},
\code{\link{Lrnr_arima}},
\code{\link{Lrnr_bartMachine}},
\code{\link{Lrnr_base}},
\code{\link{Lrnr_bilstm}},
\code{\link{Lrnr_bound}},
\code{\link{Lrnr_caret}},
\code{\link{Lrnr_cv_selector}},
\code{\link{Lrnr_cv}},
\code{\link{Lrnr_dbarts}},
\code{\link{Lrnr_define_interactions}},
\code{\link{Lrnr_density_discretize}},
\code{\link{Lrnr_density_hse}},
\code{\link{Lrnr_density_semiparametric}},
\code{\link{Lrnr_earth}},
\code{\link{Lrnr_expSmooth}},
\code{\link{Lrnr_gam}},
\code{\link{Lrnr_gbm}},
\code{\link{Lrnr_glm_fast}},
\code{\link{Lrnr_glmnet}},
\code{\link{Lrnr_glm}},
\code{\link{Lrnr_grf}},
\code{\link{Lrnr_gts}},
\code{\link{Lrnr_h2o_grid}},
\code{\link{Lrnr_hal9001}},
\code{\link{Lrnr_haldensify}},
\code{\link{Lrnr_hts}},
\code{\link{Lrnr_independent_binomial}},
\code{\link{Lrnr_lstm}},
\code{\link{Lrnr_mean}},
\code{\link{Lrnr_multiple_ts}},
\code{\link{Lrnr_multivariate}},
\code{\link{Lrnr_nnet}},
\code{\link{Lrnr_nnls}},
\code{\link{Lrnr_optim}},
\code{\link{Lrnr_pca}},
\code{\link{Lrnr_pkg_SuperLearner}},
\code{\link{Lrnr_polspline}},
\code{\link{Lrnr_pooled_hazards}},
\code{\link{Lrnr_randomForest}},
\code{\link{Lrnr_ranger}},
\code{\link{Lrnr_revere_task}},
\code{\link{Lrnr_rpart}},
\code{\link{Lrnr_rugarch}},
\code{\link{Lrnr_screener_coefs}},
\code{\link{Lrnr_screener_corP}},
\code{\link{Lrnr_screener_corRank}},
\code{\link{Lrnr_screener_randomForest}},
\code{\link{Lrnr_sl}},
\code{\link{Lrnr_solnp_density}},
\code{\link{Lrnr_solnp}},
\code{\link{Lrnr_stratified}},
\code{\link{Lrnr_subset_covariates}},
\code{\link{Lrnr_svm}},
\code{\link{Lrnr_tsDyn}},
\code{\link{Lrnr_xgboost}},
\code{\link{Pipeline}},
\code{\link{Stack}},
\code{\link{define_h2o_X}()},
\code{\link{undocumented_learner}}
}
\concept{Learners}
\keyword{data}
