% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_solnp.R
\docType{class}
\name{Lrnr_solnp}
\alias{Lrnr_solnp}
\title{Nonlinear Optimization via Augmented Lagrange}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_solnp
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This meta-learner provides fitting procedures for any pairing of loss 
function and metalearner function, subject to constraints. The
optimization problem is solved by making use of \code{Rsolnp::solnp}, using
Lagrange multipliers. For further details, consult the documentation of the
\code{Rsolnp} package.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
