% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_solnp_density.R
\docType{class}
\name{Lrnr_solnp_density}
\alias{Lrnr_solnp_density}
\title{Nonlinear Optimization via Augmented Lagrange}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_solnp_density
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This meta-learner provides fitting procedures for density estimation, finding
convex combinations of candidate density estimators by minimizing the
cross-validated negative log-likelihood loss of each candidate density. The
optimization problem is solved by making use of \code{Rsolnp::solnp}, using
Lagrange multipliers. For further details, consult the documentation of the
\code{Rsolnp} package.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}

\keyword{data}
