% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_subset_covariates.R
\docType{class}
\name{Lrnr_subset_covariates}
\alias{Lrnr_subset_covariates}
\title{Learner with Covariate Subsetting}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_subset_covariates
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides fitting procedures for subsetting covariates. It is a
convenience utility for reducing the number of covariates to be fit.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}
\keyword{data}

