% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_tsDyn.R
\docType{class}
\name{Lrnr_tsDyn}
\alias{Lrnr_tsDyn}
\title{Nonlinear Time Series Analysis}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_tsDyn
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner supports various forms of nonlinear autoregression, including
additive AR, neural nets, SETAR and LSTAR models, threshold VAR and VECM.
}
\section{Parameters}{

\describe{
\item{\code{learner}}{Available built-in time series models. Currently
available are listed by with availableModels() function.}
\item{\code{m=1}}{embedding dimension.}
\item{\code{size=1}}{number of hidden units in the neural network.}
\item{\code{lag=1}}{Number of lags to include in each regime.}
\item{\code{d=1}}{time delay.}
\item{\code{include="const"}}{Type of deterministic regressors to include.}
\item{\code{type="level"}}{Whether the variable is taken is level,
difference or a mix as in the ADF test.}
\item{\code{n.ahead="null"}}{The forecast horizon.}
\item{\code{mL=m}}{ autoregressive order for low regime.}
\item{\code{mH=m}}{ autoregressive order for high regime.}
\item{\code{mM=NULL}}{ autoregressive order for middle regime.}
\item{\code{thDelay=0}}{Delay Time delay for the threshold variable.}
\item{\code{common="none"}}{Indicates which elements are common to all
regimes.}
\item{\code{ML=seq_len(mL)}}{vector of lags for order for low.}
\item{\code{MM=NULL}}{vector of lags for order for middle.}
\item{\code{MH=seq_len(mH)}}{vector of lags for order for high.}
\item{\code{nthresh=1}}{Threshold of the model.}
\item{\code{trim=0.15}}{trimming parameter indicating the minimal
percentage of observations in each regime.}
\item{\code{sig=0.05}}{significance level for the tests to select the
number of regimes.}
\item{\code{control=list()}}{further arguments to be passed as control list
to \code{optim}.}
\item{\code{r=1}}{Number of cointegrating relationships.}
\item{\code{model="VAR"}}{Model to estimate. Choices: VAR/VECM/TAR/MTAR.}
\item{\code{I="level"}}{For VAR only: whether in the VAR variables are to
be taken in levels or as a difference.}
\item{\code{beta=NULL}}{For VECM only: imposed cointegrating value.}
\item{\code{estim="2OLS"}}{Type of estimator for the VECM (two-step
approach or Johansen MLE).}
\item{\code{exogen=NULL}}{Inclusion of exogenous variables.}
\item{\code{LRinclude="none"}}{Possibility to include in the long-run
relationship and the ECT trend.}
\item{\code{commonInter=FALSE}}{Whether the deterministic regressors are
regime specific.}
\item{\code{mTh=1}}{combination of variables with same lag order for the
transition variable.}
\item{\code{gamma=NULL}}{prespecified threshold values.}
\item{\code{dummyToBothRegimes=TRUE}}{Whether the dummy in the one
threshold model is applied to each regime. }
\item{\code{max.iter=2}}{Number of iterations for the algorithm.}
\item{\code{ngridBeta=50}}{Number of elements to search for the
cointegrating value.}
\item{\code{ngridTh=50}}{Number of elements to search for the threshold
value.}
\item{\code{th1}}{different possibilities to pre-specify an exact value, an
interval or a central point for the search of the threshold. }
\item{\code{th2}}{different possibilities to pre-specify an exact value or
a central point for the search of the second threshold. }
\item{\code{beta0=0}}{Additional regressors to include in the cointegrating
relation.}
\item{\code{...}}{Not currently used.}
}
}

\seealso{
<<<<<<< HEAD
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_bartMachine}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_bilstm}}, \code{\link{Lrnr_caret}},
  \code{\link{Lrnr_condensier}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_density_discretize}},
  \code{\link{Lrnr_density_hse}},
  \code{\link{Lrnr_density_semiparametric}},
  \code{\link{Lrnr_earth}}, \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_gam}}, \code{\link{Lrnr_gbm}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_haldensify}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_multivariate}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_optim}}, \code{\link{Lrnr_pca}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_polspline}},
  \code{\link{Lrnr_pooled_hazards}},
  \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}},
  \code{\link{Lrnr_revere_task}}, \code{\link{Lrnr_rfcde}},
  \code{\link{Lrnr_rpart}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_screener_corP}},
  \code{\link{Lrnr_screener_corRank}},
  \code{\link{Lrnr_screener_randomForest}},
  \code{\link{Lrnr_sl}}, \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}}, \code{\link{Lrnr_stratified}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_xgboost}},
  \code{\link{Pipeline}}, \code{\link{Stack}},
  \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
=======
Other Learners: 
\code{\link{Custom_chain}},
\code{\link{Lrnr_HarmonicReg}},
\code{\link{Lrnr_arima}},
\code{\link{Lrnr_bartMachine}},
\code{\link{Lrnr_base}},
\code{\link{Lrnr_bilstm}},
\code{\link{Lrnr_bound}},
\code{\link{Lrnr_caret}},
\code{\link{Lrnr_condensier}},
\code{\link{Lrnr_cv_selector}},
\code{\link{Lrnr_cv}},
\code{\link{Lrnr_dbarts}},
\code{\link{Lrnr_define_interactions}},
\code{\link{Lrnr_density_discretize}},
\code{\link{Lrnr_density_hse}},
\code{\link{Lrnr_density_semiparametric}},
\code{\link{Lrnr_earth}},
\code{\link{Lrnr_expSmooth}},
\code{\link{Lrnr_gam}},
\code{\link{Lrnr_gbm}},
\code{\link{Lrnr_glm_fast}},
\code{\link{Lrnr_glmnet}},
\code{\link{Lrnr_glm}},
\code{\link{Lrnr_grf}},
\code{\link{Lrnr_h2o_grid}},
\code{\link{Lrnr_hal9001}},
\code{\link{Lrnr_haldensify}},
\code{\link{Lrnr_independent_binomial}},
\code{\link{Lrnr_lstm}},
\code{\link{Lrnr_mean}},
\code{\link{Lrnr_multivariate}},
\code{\link{Lrnr_nnls}},
\code{\link{Lrnr_optim}},
\code{\link{Lrnr_pca}},
\code{\link{Lrnr_pkg_SuperLearner}},
\code{\link{Lrnr_polspline}},
\code{\link{Lrnr_pooled_hazards}},
\code{\link{Lrnr_randomForest}},
\code{\link{Lrnr_ranger}},
\code{\link{Lrnr_revere_task}},
\code{\link{Lrnr_rfcde}},
\code{\link{Lrnr_rpart}},
\code{\link{Lrnr_rugarch}},
\code{\link{Lrnr_screener_corP}},
\code{\link{Lrnr_screener_corRank}},
\code{\link{Lrnr_screener_randomForest}},
\code{\link{Lrnr_sl}},
\code{\link{Lrnr_solnp_density}},
\code{\link{Lrnr_solnp}},
\code{\link{Lrnr_stratified}},
\code{\link{Lrnr_subset_covariates}},
\code{\link{Lrnr_svm}},
\code{\link{Lrnr_xgboost}},
\code{\link{Pipeline}},
\code{\link{Stack}},
\code{\link{define_h2o_X}()},
\code{\link{undocumented_learner}}
>>>>>>> 63d08007057d111e9839e207e70488ebc006f5ad
}
\concept{Learners}
\keyword{data}
\section{Super class}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{Lrnr_tsDyn}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Lrnr_tsDyn$new()}}
\item \href{#method-clone}{\code{Lrnr_tsDyn$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained">}\href{../../sl3/html/Lrnr_base.html#method-assert_trained}{\code{sl3::Lrnr_base$assert_trained()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain">}\href{../../sl3/html/Lrnr_base.html#method-base_chain}{\code{sl3::Lrnr_base$base_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict">}\href{../../sl3/html/Lrnr_base.html#method-base_predict}{\code{sl3::Lrnr_base$base_predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_train">}\href{../../sl3/html/Lrnr_base.html#method-base_train}{\code{sl3::Lrnr_base$base_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain">}\href{../../sl3/html/Lrnr_base.html#method-chain}{\code{sl3::Lrnr_base$chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain">}\href{../../sl3/html/Lrnr_base.html#method-custom_chain}{\code{sl3::Lrnr_base$custom_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_range}{\code{sl3::Lrnr_base$get_outcome_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_type}{\code{sl3::Lrnr_base$get_outcome_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict">}\href{../../sl3/html/Lrnr_base.html#method-predict}{\code{sl3::Lrnr_base$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold">}\href{../../sl3/html/Lrnr_base.html#method-predict_fold}{\code{sl3::Lrnr_base$predict_fold()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="print">}\href{../../sl3/html/Lrnr_base.html#method-print}{\code{sl3::Lrnr_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="sample">}\href{../../sl3/html/Lrnr_base.html#method-sample}{\code{sl3::Lrnr_base$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train">}\href{../../sl3/html/Lrnr_base.html#method-set_train}{\code{sl3::Lrnr_base$set_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates">}\href{../../sl3/html/Lrnr_base.html#method-subset_covariates}{\code{sl3::Lrnr_base$subset_covariates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train">}\href{../../sl3/html/Lrnr_base.html#method-train}{\code{sl3::Lrnr_base$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners">}\href{../../sl3/html/Lrnr_base.html#method-train_sublearners}{\code{sl3::Lrnr_base$train_sublearners()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_tsDyn$new(
  learner,
  m = 1,
  size = 1,
  lag = 1,
  d = 1,
  include = "const",
  type = "level",
  n.ahead = NULL,
  mL = m,
  mH = m,
  mM = NULL,
  thDelay = 0,
  common = "none",
  ML = seq_len(mL),
  MM = NULL,
  MH = seq_len(mH),
  nthresh = 1,
  trim = 0.15,
  sig = 0.05,
  control = list(),
  r = 1,
  model = "VAR",
  I = "level",
  beta = NULL,
  estim = "2OLS",
  exogen = NULL,
  LRinclude = "none",
  commonInter = FALSE,
  mTh = 1,
  gamma = NULL,
  dummyToBothRegimes = TRUE,
  max.iter = 2,
  ngridBeta = 50,
  ngridTh = 50,
  th1 = list(exact = NULL, int = c("from", "to"), around = "val"),
  th2 = list(exact = NULL, int = c("from", "to"), around = "val"),
  beta0 = 0,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_tsDyn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
