% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_base.R
\name{make_learner}
\alias{make_learner}
\title{Instantiate a Learner}
\usage{
make_learner(learner_class, ...)
}
\arguments{
\item{learner_class}{the learner class to instantiate}

\item{...}{parameters with which to instantiate the learner}
}
\description{
Instantiate a Learner
}
