% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\docType{class}
\name{Pipeline}
\alias{Pipeline}
\title{Pipeline (chain) of learners.
A Pipeline of learners is a way to "chain" Learners together, where the output of one learner is used as output for the next learner. This can be used for things like screening, two stage machine learning methods, and Super Learning. A pipeline is fit by fitting the first \code{Learner}, calling \code{chain()} to create the next task, which becomes the training data for the next \code{Learner}. Similarly, for prediction, the predictions from the first \code{Learner} become the data to predict on for the next \code{Learner}.}
\format{\code{\link{R6Class}} object.}
\usage{
Pipeline
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Pipeline (chain) of learners.
A Pipeline of learners is a way to "chain" Learners together, where the output of one learner is used as output for the next learner. This can be used for things like screening, two stage machine learning methods, and Super Learning. A pipeline is fit by fitting the first \code{Learner}, calling \code{chain()} to create the next task, which becomes the training data for the next \code{Learner}. Similarly, for prediction, the predictions from the first \code{Learner} become the data to predict on for the next \code{Learner}.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of learners to chain.}
}}

\section{Methods}{

\describe{
\item{\code{new(...)}}{This method is used to create a pipeline of learners. Arguments should be indiviual \code{Learner}s, in the order they should be applied.}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_sl}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Stack}}
}
\concept{Learners}
\keyword{data}
