% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_utils.R
\name{pooled_hazard_task}
\alias{pooled_hazard_task}
\title{Generate A Pooled Hazards Task from a Failure Time (or Categorical) Task}
\usage{
pooled_hazard_task(task)
}
\arguments{
\item{task}{a task where the outcome is failure time}
}
\description{
Generate A Pooled Hazards Task from a Failure Time (or Categorical) Task
}
