% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_helpers.R
\name{predict_classes}
\alias{predict_classes}
\title{Predict Class from Predicted Probabilities}
\usage{
predict_classes(predictions)
}
\arguments{
\item{predictions}{the nxc matrix where each row are predicted probabilities for
one observation for each of c classes}
}
\value{
a vector of length n, the predicted class labels as a factor variable
}
\description{
Returns the most likely class label for each row of predicted class probabilities
}
