% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{prediction_plot}
\alias{prediction_plot}
\title{Plot predicted and true values for diganostic purposes}
\usage{
prediction_plot(learner_fit)
}
\arguments{
\item{learner_fit}{A fit sl3 learner object. Ideally from a Lrnr_sl}
}
\value{
A ggplot2 object
}
\description{
If a Lrnr_sl fit is provided, predictions will be generated from the cross-validated learner fits and final metalearner fit.
Otherwise, non cross-validated predictions will be used an an error will be thrown
}
