% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_fit_test}
\alias{reduce_fit_test}
\title{Drop components from learner fits}
\usage{
reduce_fit_test(learner_fit)
}
\arguments{
\item{learner_fit}{A \code{Lrnr_*} object, inheriting from \code{Lrnr_base},
that has been trained.}
}
\description{
Given learner fit, sequentially drop components from internal fit object,
keeping track of which components are needed for prediction.
}
\keyword{internal}
