% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_add_user_args}
\alias{replace_add_user_args}
\title{Replace an argument in \code{mainArgs} if it also appears in \code{userArgs}.
Add any argument from \code{userArgs} that appears in \code{formals(fun)} of
the function \code{fun}.}
\usage{
replace_add_user_args(mainArgs, userArgs, fun)
}
\description{
Replace an argument in \code{mainArgs} if it also appears in \code{userArgs}.
Add any argument from \code{userArgs} that appears in \code{formals(fun)} of
the function \code{fun}.
}
\keyword{internal}
