% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_dim}
\alias{safe_dim}
\title{dim that works for vectors too}
\usage{
safe_dim(x)
}
\arguments{
\item{x}{the object to get dimensions from}
}
\description{
\code{safe_dim} tries to get dimensions from \code{dim} and falls back on
\code{length} if \code{dim} returns \code{NULL}
}
