% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3_Task.R
\name{sl3_Task$new}
\alias{sl3_Task$new}
\title{Create a new sl3_Task}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the underlying data
#' \describe{
\item{\code{new(data, covariates, outcome = NULL, outcome_type = NULL, outcome_levels = NULL,
              id = NULL, weights = NULL, offset = NULL, nodes = NULL, column_names = NULL,
              row_index = NULL, folds = NULL)}}{
This method is used to create an object of this class.
\describe{
\item{\code{data}}{A \code{data.frame} or \code{data.table} containing the raw data}
}}
\item{\code{next_in_chain(new_X)}}{
Generates a copy of this task with the set of covariates redefined. This is mostly to be used internally for \code{\link{Pipeline}}s}
}}
}
\description{
Create a new sl3_Task
}
