% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sl3Options}
\alias{sl3Options}
\title{Querying/setting a single \code{sl3} option}
\usage{
sl3Options(o, value)
}
\arguments{
\item{o}{Option name (string).}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{sl3} options, just run this function without any parameters
 provided. To query only one value, pass the first parameter. To set that,
 use the \code{value} parameter too.
}
\examples{
\dontrun{
sl3Options()
sl3Options('sl3.verbose')
sl3Options('sl3.temp.dir')
sl3Options('sl3.verbose', TRUE)
}
}
