% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack.R
\docType{class}
\name{Stack}
\alias{Stack}
\title{Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns. Currently, train fits the learners one-at-a-time, but this will be parallelized with \link{future} going forward.}
\format{\code{\link{R6Class}} object.}
\usage{
Stack
}
\value{
\code{\link{Learner}} object with methods for training and prediction
}
\description{
Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns. Currently, train fits the learners one-at-a-time, but this will be parallelized with \link{future} going forward.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of learners to stack}
}}

\section{Methods}{

\describe{
  \item{\code{new(...)}}{This method is used to create a stack of learners. Arguments should be indiviual \code{Learner}s.}
  }
}

\seealso{
Other Learners: \code{\link{CV_Learner}},
  \code{\link{GLM_Learner}}, \code{\link{Learner}},
  \code{\link{Pipeline}}
}
\keyword{data}
