% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_dt_cols}
\alias{subset_dt_cols}
\title{Subset data.table columns}
\usage{
subset_dt_cols(dt, cols)
}
\arguments{
\item{dt}{A \code{\link[data.table]{data.table}}.}

\item{cols}{Columns to subset.}
}
\value{
Subset of input \code{\link[data.table]{data.table}}.
}
\description{
Subset data.table columns
}
\keyword{internal}
