% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{transform_losses}
\alias{transform_losses}
\title{Transform losses for loss functions with transform attribute}
\usage{
transform_losses(losses, type)
}
\arguments{
\item{losses}{Numeric vector of losses returned by a loss function}

\item{type}{A valid transformation type.}
}
\description{
Transform losses for loss functions with transform attribute
}
