% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{transform_risk}
\alias{transform_risk}
\title{Transform risks for a risk function with transform attribute}
\usage{
transform_risk(risk, type)
}
\arguments{
\item{risk}{Numeric vector or scalar of risk(s) returned by a risk function}

\item{type}{A valid transformation type}
}
\description{
Transform risks for a risk function with transform attribute
}
