% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{true_obj_size}
\alias{true_obj_size}
\title{Estimate object size using serialization}
\usage{
true_obj_size(obj)
}
\arguments{
\item{obj}{the object to get the size of}
}
\value{
the size of \code{obj} in bytes
}
\description{
Attempts to get a better estimate of object size than that returned by \code{\link{object.size}}
}
