% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_Learner.R, R/Model_Matrix.R,
%   R/SL_Learner.R, R/SL_Screener.R, R/Super_Learner.R, R/fastGLM_Learner.R,
%   R/h2oGLM_Learner.R, R/sl3-package.R
\docType{data}
\name{GLM_Learner}
\alias{GLM_Learner}
\alias{Model_Matrix}
\alias{SL_Learner}
\alias{SL_Screener}
\alias{Super_Learner}
\alias{fastGLM_Learner}
\alias{h2oGLM_Learner}
\alias{undocumented_learner}
\title{Undocumented Learner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GLM_Learner

Model_Matrix

SL_Learner

SL_Screener

Super_Learner

fastGLM_Learner

h2oGLM_Learner
}
\value{
\code{\link{Learner}} object with methods for training and prediction
}
\description{
We haven't documented this one yet.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}
}}

\seealso{
Other Learners: \code{\link{CV_Learner}},
  \code{\link{Learner}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
\keyword{datasets}
