% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_to_list}
\alias{args_to_list}
\title{Get all arguments of parent call (both specified and defaults) as list}
\usage{
args_to_list()
}
\value{
A \code{list} of all arguments for the parent function call.
}
\description{
Get all arguments of parent call (both specified and defaults) as list
}
