% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Truncates predictions to ensure loss function is bounded.}
\usage{
bound(preds, bounds = 0.001)
}
\arguments{
\item{preds}{A \code{numeric} vector of predictions to to be bounded.}

\item{bounds}{Either a \code{numeric} vector of length two, giving the
closed interval (lower, upper), or just a lower bound. In the latter case,
the upper bound is computed as (1 - lower). The default is 0.001.}
}
\value{
Truncated predictions.
}
\description{
Truncates predictions to ensure loss function is bounded.
}
\keyword{internal}
