% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3-package.R
\docType{data}
\name{cpp}
\alias{cpp}
\title{Subset of growth data from the collaborative perinatal project (CPP)}
\source{
\url{https://catalog.archives.gov/id/606622}

Broman, Sarah. "The collaborative perinatal project: an overview." Handbook of longitudinal research 1 (1984): 185-227.
}
\usage{
cpp
}
\description{
Subset of growth data from the collaborative perinatal project (CPP).
}
\examples{
data(cpp)
head(cpp)
}
\keyword{data}
