% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_chain.R
\docType{class}
\name{Custom_chain}
\alias{Custom_chain}
\alias{customize_chain}
\title{Customize chaining for a learner}
\format{\code{\link{R6Class}} object.}
\usage{
Custom_chain

customize_chain(learner, chain_fun)
}
\arguments{
\item{learner}{A \code{sl3} learner to modify.}

\item{chain_fun}{A function with arguments \code{learner} and \code{task}
that defines the new chain behavior.}
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
prediction
}
\description{
Customize chaining for a learner

This function wraps a learner in such a way that the behavior of
\code{learner$chain} is modified to use a new function definition.
\code{learner$train} and \code{learner$predict} are unaffected.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of learners to chain.}
}}

\section{Methods}{

\describe{
\item{\code{new(...)}}{This method is used to create a pipeline of
learners. Arguments should be indiviual \code{Learner}s, in the order they
should be applied.}
}
}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_bartMachine}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_bilstm}},
  \code{\link{Lrnr_condensier}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pca}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}}, \code{\link{Lrnr_rpart}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Lrnr_xgboost}}, \code{\link{Pipeline}},
  \code{\link{Stack}}, \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
