% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Lrnr_sl.R
\name{CV_lrnr_sl}
\alias{CV_lrnr_sl}
\title{Estimate Cross-Validated Risk of Super Learner}
\usage{
CV_lrnr_sl(lrnr_sl, task, loss_fun)
}
\arguments{
\item{lrnr_sl}{a Lrnr_sl object specifying the Super Learner}

\item{task}{the task used for training and performance assessment}

\item{loss_fun}{the loss function used to evaluate Super Learner}
}
\description{
Estimate Cross-Validated Risk of Super Learner
}
