% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Lrnr_sl.R
\name{CV_lrnr_sl}
\alias{CV_lrnr_sl}
\title{Estimates cross-validated risk of the Super Learner}
\usage{
CV_lrnr_sl(lrnr_sl, task, eval_fun)
}
\arguments{
\item{lrnr_sl}{a \code{Lrnr_sl} object specifying the Super Learner.}

\item{task}{the task used for training and performance assessment.}

\item{eval_fun}{the evaluation function, either a loss or risk function, for
evaluating the Super Learner's predictions.}
}
\description{
Estimates cross-validated risk of the Super Learner
}
