% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_metalearner.R
\name{default_metalearner}
\alias{default_metalearner}
\title{Automatically Defined Metalearner}
\usage{
default_metalearner(outcome_type)
}
\arguments{
\item{outcome_type}{a Variable_Type object}
}
\description{
A sensible metalearner is chosen based on the outcome type.
}
\details{
For binary and continuous outcome types, the default metalearner is
non-negative least squares (NNLS) regression (\code{\link{Lrnr_nnls}}), and
for others the metalearner is \code{\link{Lrnr_solnp}} with an appropriate
loss and combination function, shown in the table below.\tabular{lll}{
   Outcome Type \tab Combination Function \tab Loss Function \cr
   categorical \tab metalearner_linear_multinomial \tab loss_loglik_multinomial \cr
   multivariate \tab metalearner_linear_multivariate \tab loss_squared_error_multivariate \cr
}
}
