% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_variable_type}
\alias{guess_variable_type}
\title{guess variable type}
\usage{
guess_variable_type(x, pcontinuous = 0.2)
}
\arguments{
\item{x}{the variable}

\item{pcontinuous}{The proportion of observations that have to be unique before a variable is determined to be continuous}
}
\description{
guess variable type
}
