% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_randomForest.R, R/importance.R
\name{importance}
\alias{importance}
\title{Importance
Extract variable importance measures produced by
\code{\link[randomForest]{randomForest}} and order in decreasing order of
importance.}
\usage{
importance(fit, loss = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"))

importance(fit, loss = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"))
}
\arguments{
\item{fit}{A trained cross-validated learner (e.g., cv stack, super learner),
from which cross-validated predictions can be generated.}

\item{loss}{The loss function for evaluating the risk. Defaults according to
outcome type: squared error loss for continuous outcomes, and negative
log-likelihood loss for discrete outcomes. See \code{\link{loss_functions}}.}

\item{fold_number}{The fold number to use for obtaining the predictions
from the fit. Either a positive integer for obtaining predictions from a
specific fold's fit; \code{"full"} for obtaining predictions from a fit on
all of the data, or \code{"validation"} (default) for obtaining
cross-validated predictions, where the data used for training and prediction
never overlaps across the folds. Note that if a positive integer or
\code{"full"} is supplied, then there will be overlap between the  data
used for training and prediction.}

\item{type}{Which method should be used to obscure the relationship between
the covariate and the outcome. When type is \code{"remove"} (default), each
covariate is removed from the task and the cross-validated learner is refit
to this modified task and then predictions are obtained from this refit.
When type is \code{"permute"}, each covariate is permuted (sampled without
replacement), and then predictions are obtained from this modified data with
one permuted covariate.}

\item{importance_metric}{Either \code{"ratio"} (default), which returns
the risk with the permuted/removed X divided by observed risk with all X; or
\code{"difference"}, which returns the difference between the risk with the
permuted/removed X and the observed risk.}
}
\value{
A \code{data.table} of variable importance for each covariate.
}
\description{
Function that takes a cross-validated fit (i.e., cross-validated learner that
has already been trained on a task), which could be a cross-validated single
learner or super learner, to generate a loss-based variable importance
measure for each predictor, where the predictors are the covariates in the
trained task. This function generates a \code{data.table} in which each row
corresponds to the risk difference or risk ratio between the following
two risks: the risk when a predictor is permuted or removed, and the original
risk (i.e., when all predictors are included). A higher risk ratio/difference
corresponds to a more important predictor. A plot can be generated from the
returned \code{data.table} by calling companion function
\code{plot_importance}.
}
\keyword{importance}
\keyword{variable}
