% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_randomForest.R, R/importance.R
\name{importance}
\alias{importance}
\title{Importance
Extract variable importance measures produced by
\code{\link[randomForest]{randomForest}} and order in decreasing order of
importance.}
\usage{
importance(fit, eval_fun = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"), covariate_groups = NULL, B = 100L,
  cores = parallel::detectCores() - 1)

importance(fit, eval_fun = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"), covariate_groups = NULL, B = 100L,
  cores = parallel::detectCores() - 1)
}
\arguments{
\item{fit}{A trained cross-validated learner (e.g., cv stack, super learner),
from which cross-validated predictions can be generated.}

\item{eval_fun}{The evaluation function (risk or loss function) for
evaluating the risk. Defaults vary based on the outcome type: squared error
loss for continuous outcomes, and negative log-likelihood loss for discrete
outcomes. See \code{\link{loss_functions}} and \code{\link{risk_functions}}
for options.}

\item{fold_number}{The fold number to use for obtaining the predictions
from the fit. Either a positive integer for obtaining predictions from a
specific fold's fit; \code{"full"} for obtaining predictions from a fit on
all of the data, or \code{"validation"} (default) for obtaining
cross-validated predictions, where the data used for training and prediction
never overlaps across the folds. Note that if a positive integer or
\code{"full"} is supplied, then there will be overlap between the  data
used for training and prediction.}

\item{type}{Which method should be used to obscure the relationship between
the covariate and the outcome. When type is \code{"remove"} (default), each
covariate is removed from the task and the cross-validated learner is refit
to this modified task and then predictions are obtained from this refit.
When type is \code{"permute"}, each covariate is permuted (sampled without
replacement), and then predictions are obtained from this modified data with
one permuted covariate.}

\item{importance_metric}{Either \code{"ratio"} (default), which returns
the risk with the permuted/removed X divided by observed risk with all X; or
\code{"difference"}, which returns the difference between the risk with the
permuted/removed X and the observed risk.}

\item{covariate_groups}{Optional named list covariate groups which will
invoke variable importance evaluation at the group-level, by
removing/permuting all covariates in the same group together. If covariates
in the task are not specified in the list of groups, then those covariates
will be added as additional single-covariate groups.}

\item{B}{Optional integer specifying the number of permutations to average
the permuted risk over when \code{type = "permute"}.}

\item{cores}{Number of cores for parallelizing the \code{B} permutations.}
}
\value{
A \code{data.table} of variable importance for each covariate.
}
\description{
Function that takes a cross-validated fit (i.e., cross-validated learner that
has already been trained on a task), which could be a cross-validated single
learner or super learner, to generate a risk-based variable importance
measure for each covariate in the trained task, or for each group of
covariates. This function outputs a \code{data.table}, where each row
corresponds to the risk difference or the risk ratio between the following
two risks: the risk when a covariate (or group of covariates) is permuted or
removed, and the original risk (i.e., when all covariates are included as
they were in the observed data). A higher risk ratio/difference corresponds
to a more important covariate/group. A plot can be generated from the
returned \code{data.table} by calling companion function
\code{plot_importance}.
}
\keyword{importance}
\keyword{variable}
