% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_helpers.R
\name{delayed_make_learner}
\alias{delayed_make_learner}
\alias{learner_train}
\alias{delayed_learner_train}
\alias{learner_fit_predict}
\alias{delayed_learner_fit_predict}
\alias{learner_fit_chain}
\alias{delayed_learner_fit_chain}
\alias{learner_subset_covariates}
\alias{learner_process_formula}
\alias{delayed_learner_subset_covariates}
\alias{delayed_learner_process_formula}
\title{Learner helpers}
\usage{
delayed_make_learner(learner_class, ...)

learner_train(learner, task, trained_sublearners)

delayed_learner_train(learner, task, name = NULL)

learner_fit_predict(learner_fit, task = NULL)

delayed_learner_fit_predict(learner_fit, task = NULL)

learner_fit_chain(learner_fit, task = NULL)

delayed_learner_fit_chain(learner_fit, task = NULL)

learner_subset_covariates(learner, task)

learner_process_formula(learner, task)

delayed_learner_subset_covariates(learner, task)

delayed_learner_process_formula(learner, task)
}
\arguments{
\item{learner_class}{The learner class to instantiate.}

\item{...}{Parameters with which to instantiate the learner.}

\item{learner}{A learner object to fit to the task.}

\item{task}{The task on which to fit.}

\item{trained_sublearners}{Any data obtained from a \code{train_sublearners}
step.}

\item{name}{a more detailed name for this delayed task, if necessary}

\item{learner_fit}{a learner object that has already been fit}
}
\description{
Learner helpers
}
