% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_functions}
\alias{loss_functions}
\alias{loss_squared_error}
\alias{loss_loglik_true_cat}
\alias{loss_loglik_binomial}
\alias{loss_loglik_multinomial}
\alias{loss_squared_error_multivariate}
\title{Loss Function Definitions}
\usage{
loss_squared_error(pred, observed)

loss_loglik_true_cat(pred, observed)

loss_loglik_binomial(pred, observed)

loss_loglik_multinomial(pred, observed)

loss_squared_error_multivariate(pred, observed)
}
\arguments{
\item{pred}{A vector of predicted values}

\item{observed}{A vector of observed values}
}
\value{
A vector of loss values
}
\description{
Loss functions for use in evaluating learner fits
}
