% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_base.R
\docType{class}
\name{Lrnr_base}
\alias{Lrnr_base}
\title{Base Class for all sl3 learners.}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_base
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Generally this base Learner class shouldn't be instantiated. It's intended to be an abstract class, although abstract classes aren't explicitly supported by R6.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}

\item{\code{is_trained}}{A boolean indicating that the learner has been already trained.}

\item{\code{name}}{A short name used to reference predictions from this class}
}}

\section{Methods}{

\describe{
  \item{\code{new(...)}}{This method is used to create an object of this class.}
  \item{\code{train(task)}}{Trains this Learner on the data in the \code{task} object. Returns a new \code{Learner} object containing the fit of this Learner to the data in \code{task}. This can then be used with the \code{predict} and \code{chain} methods.}
  \item{\code{predict(task=NULL)}}{When called from a \code{Learner} object generated by a call to \code{train}, can be used to generate predictions on the data in \code{task}. If no \code{task} is specified, the task used to train the Learner will be used.}
  \item{\code{chain(task=NULL)}}{When called from a \code{Learner} object generated by a call to \code{train}, will generate a new task based on the pre be used to generate predictions on the data in \code{task}. This can be used to train a new Learner on the output of this one. This is mostly accomplished by using \code{\link{Pipeline}} Learner. If no \code{task} is specified, the task used to train the Learner will be used.}
  }
}

\seealso{
Other Learners: \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_glm}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
