% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_caret.R
\docType{class}
\name{Lrnr_caret}
\alias{Lrnr_caret}
\title{Wrapping Learner for Package Caret}
\format{
\code{\link{R6Class}} object.
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner wraps \code{\link[caret]{train}}, providing a procedure to fit the
algorithms available in package \code{caret}.
}
\section{Parameters}{

\describe{
\item{\code{algorithm}}{The algorithm to use. This argument is passed as
\code{method} to \code{\link[caret]{train}}. For a detailed description of the
methods available in the package, please consult the documentation for
\code{\link[caret]{train}}.}

\item{\code{method}}{An optional argument specifying the method for
parameter selection criterion. Default is set to "CV" instead of the
\code{\link[caret]{train}} default "boot". This argument is passed to
\code{trControl} and will not be effective if \code{trControl} is specified. For a
detailed description, please consult the documentation for
\code{\link[caret]{train}}.}

\item{\code{metric}}{An optional argument specifying the summary metric to
be used to select the optimal model. If not specified, it will be inferred
depending on the outcome type. For a detailed description, please consult
the documentation for \code{\link[caret]{train}}.}

\item{\code{trControl}}{An optional \code{\link[caret]{trainControl}}
object controlling the computational nuances. If not specified, it will be
initialized with \code{method}. For a detailed description, please consult the
documentation for \code{\link[caret]{train}}.}

\item{\code{...}}{Other parameters passed to \code{\link[caret]{train}}.}
}
}

\seealso{
Other Learners: 
\code{\link{Custom_chain}},
\code{\link{Lrnr_HarmonicReg}},
\code{\link{Lrnr_arima}},
\code{\link{Lrnr_bartMachine}},
\code{\link{Lrnr_base}},
\code{\link{Lrnr_bilstm}},
\code{\link{Lrnr_bound}},
\code{\link{Lrnr_cv_selector}},
\code{\link{Lrnr_cv}},
\code{\link{Lrnr_dbarts}},
\code{\link{Lrnr_define_interactions}},
\code{\link{Lrnr_density_discretize}},
\code{\link{Lrnr_density_hse}},
\code{\link{Lrnr_density_semiparametric}},
\code{\link{Lrnr_earth}},
\code{\link{Lrnr_expSmooth}},
\code{\link{Lrnr_gam}},
\code{\link{Lrnr_gbm}},
\code{\link{Lrnr_glm_fast}},
\code{\link{Lrnr_glmnet}},
\code{\link{Lrnr_glm}},
\code{\link{Lrnr_grf}},
\code{\link{Lrnr_gts}},
\code{\link{Lrnr_h2o_grid}},
\code{\link{Lrnr_hal9001}},
\code{\link{Lrnr_haldensify}},
\code{\link{Lrnr_hts}},
\code{\link{Lrnr_independent_binomial}},
\code{\link{Lrnr_lstm}},
\code{\link{Lrnr_mean}},
\code{\link{Lrnr_multiple_ts}},
\code{\link{Lrnr_multivariate}},
\code{\link{Lrnr_nnet}},
\code{\link{Lrnr_nnls}},
\code{\link{Lrnr_optim}},
\code{\link{Lrnr_pca}},
\code{\link{Lrnr_pkg_SuperLearner}},
\code{\link{Lrnr_polspline}},
\code{\link{Lrnr_pooled_hazards}},
\code{\link{Lrnr_randomForest}},
\code{\link{Lrnr_ranger}},
\code{\link{Lrnr_revere_task}},
\code{\link{Lrnr_rpart}},
\code{\link{Lrnr_rugarch}},
\code{\link{Lrnr_screener_augment}},
\code{\link{Lrnr_screener_coefs}},
\code{\link{Lrnr_screener_correlation}},
\code{\link{Lrnr_screener_importance}},
\code{\link{Lrnr_sl}},
\code{\link{Lrnr_solnp_density}},
\code{\link{Lrnr_solnp}},
\code{\link{Lrnr_stratified}},
\code{\link{Lrnr_subset_covariates}},
\code{\link{Lrnr_svm}},
\code{\link{Lrnr_tsDyn}},
\code{\link{Lrnr_xgboost}},
\code{\link{Pipeline}},
\code{\link{Stack}},
\code{\link{define_h2o_X}()},
\code{\link{undocumented_learner}}
}
\concept{Learners}
\keyword{data}
