% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_condensier.R
\docType{class}
\name{Lrnr_condensier}
\alias{Lrnr_condensier}
\title{Conditional Density Estimation}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_condensier
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides facilities for conditional density estimation using
the \code{condensier} package. Fitting is done with the
\code{\link[condensier]{fit_density}} function.
}
\section{Parameters}{

\describe{
\item{\code{bin_method=c("equal.mass", "equal.len", "dhist")}}{The type of
smoothing to be performed. See documentation of the \code{condensier}
package for details.}
\item{\code{nbins=5}}{The number of observations per bin. See documentation
of the \code{condensier} package for details.}
\item{\code{max_n_cat = 20}}{Maximum number of unique levels for categorical
outcomes. See documentation of the \code{condensier} package for details.}
\item{\code{pool = FALSE}}{Whether pooling of data across bins should be
performed. See documentation of the \code{condensier} package for details.}
\item{\code{max_n_bin=NA}}{Maximum number of observations per single bin for
continuous outcome. See documentation of the \code{condensier} package for
details.}
\item{\code{parfit=FALSE}}{Whether to invoke parallelization in the fitting
procedure. See documentation of the \code{condensier} package for details.}
\item{\code{bin_estimator=make_learner(Lrnr_glm_fast, family=binomial())}}{
The classification algorithm to be used in the fitting process. See
documentation of the \code{condensier} package for details.}
\item{\code{intrvls=NULL}}{An interval range to be used for custom bin
definitions. See documentation of the \code{condensier} package for
details.}
}
}

\section{Common Parameters}{


Individual learners have their own sets of parameters. Below is a list of shared parameters, implemented by \code{Lrnr_base}, and shared
by all learners.

\describe{
  \item{\code{covariates}}{A character vector of covariates. The learner will use this to subset the covariates for any specified task}
  \item{\code{outcome_type}}{A \code{\link{variable_type}} object used to control the outcome_type used by the learner. Overrides the task outcome_type if specified}
  \item{\code{...}}{All other parameters should be handled by the invidual learner classes. See the documentation for the learner class you're instantiating}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_bartMachine}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_bilstm}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pca}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}}, \code{\link{Lrnr_rpart}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Lrnr_xgboost}}, \code{\link{Pipeline}},
  \code{\link{Stack}}, \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
