% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_cv.R
\docType{class}
\name{Lrnr_cv}
\alias{Lrnr_cv}
\title{Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_cv
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Fit/Predict a learner with CV
A wrapper around any learner that generates cross-validate predictions
}
\section{Fields}{

\describe{
\item{\code{params}}{A learner to be wrapped}
}}

\section{Methods}{

\describe{
  \item{\code{new(learner)}}{Wraps the \code{learner} in a cross-validation layer.}
  }
}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_expSmooth}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
