% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_expSmooth.R
\docType{class}
\name{Lrnr_expSmooth}
\alias{Lrnr_expSmooth}
\title{Exponential Smoothing}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_expSmooth
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner supports exponential smoothing models.
}
\section{Fields}{

\describe{
\item{\code{model}}{Three-character string identifying method. In all cases, "N"=none, "A"=additive,
"M"=multiplicative, and "Z"=automatically selected. The first letter denotes the error type, second
letter denotes the trend type, third letter denotes the season type. For example, "ANN" is simple
exponential smoothing with additive errors, "MAM" is multiplicative Holt-Winters' methods with
multiplicative errors, etc.}

\item{\code{damped}}{If TRUE, use a damped trend (either additive or multiplicative). If NULL, both damped and non-damped trends will be tried and the best model (according to the information criterion ic) returned.}

\item{\code{alpha}}{Value of alpha. If NULL, it is estimated.}

\item{\code{beta}}{Value of beta. If NULL, it is estimated.}

\item{\code{gamma}}{Value of gamma. If NULL, it is estimated.}

\item{\code{phi}}{Value of phi. If NULL, it is estimated.}

\item{\code{lambda}}{Box-Cox transformation parameter. Ignored if NULL. When lambda is specified, additive.only is set to TRUE.}

\item{\code{additive.only}}{If TRUE, will only consider additive models.}

\item{\code{biasadj}}{Use adjusted back-transformed mean for Box-Cox transformations.}

\item{\code{lower}}{Lower bounds for the parameters (alpha, beta, gamma, phi)}

\item{\code{upper}}{Upper bounds for the parameters (alpha, beta, gamma, phi)}

\item{\code{opt.crit}}{Optimization criterion.}

\item{\code{nmse}}{Number of steps for average multistep MSE (1<=nmse<=30).}

\item{\code{bounds}}{Type of parameter space to impose: "usual" indicates all parameters must lie between specified lower and upper bounds; "admissible" indicates parameters must lie in the admissible space; "both" (default) takes the intersection of these regions.}

\item{\code{ic}}{Information criterion to be used in model selection.}

\item{\code{restrict}}{If TRUE, models with infinite variance will not be allowed.}

\item{\code{allow.multiplicative.trend}}{If TRUE, models with multiplicative trend are allowed when searching for a model.}

\item{\code{use.initial.values}}{If TRUE and model is of class "ets", then the initial values in the model are also not re-estimated.}

\item{\code{n.ahead}}{The forecast horizon. If not specified, returns forecast of size \code{task$X}.}

\item{\code{freq}}{the number of observations per unit of time.}
}}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
