% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_glmnet.R
\docType{class}
\name{Lrnr_glmnet}
\alias{Lrnr_glmnet}
\title{Random Forest Models}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_glmnet
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides fitting procedures for elastic net models, using the
\code{glmnet} package. For details on the fitting procedure, consult
the documentation of the \code{glmnet} package.
}
\section{Fields}{

\describe{
\item{\code{...}}{Additional arguments. Currently unused.}
}}
\keyword{data}

