% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_h2o_glm.R
\docType{class}
\name{Lrnr_h2o_glm}
\alias{Lrnr_h2o_glm}
\alias{define_h2o_X}
\title{h2o Model Definition}
\format{\code{\link{R6Class}} object.}
\usage{
define_h2o_X(task)

Lrnr_h2o_glm
}
\arguments{
\item{task}{An object of type \code{Lrnr_base} as defined in this package.}
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
Definition of \code{h2o} type models. This function is for internal use only.
This function uploads input data into an \code{h2o.Frame}, allowing the data
to be subset to the \code{task$X} \code{data.table} by a smaller set of
covariates if spec'ed in params.

This learner provides faster fitting procedures for generalized linear models
by using the \code{h2o} package. Such procedures use the H2O platform to fit
GLMs in a computationally efficient manner. For details on the procedure,
consult the documentation of the \code{h2o} package.
}
\section{Fields}{

\describe{
\item{\code{family}}{A character describing the type of model to be fit in terms of
the error family of the GLM.}

\item{\code{covariates}}{Further variables to used in fitting GLMs with \code{h2o}.}

\item{\code{...}}{Additional arguments.}
}}
\keyword{data}

