% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_h2o_grid.R
\docType{class}
\name{Lrnr_h2o_grid}
\alias{Lrnr_h2o_grid}
\alias{Lrnr_h2o_classifier}
\alias{Lrnr_h2o_mutator}
\title{Grid Search Models with h2o}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_h2o_grid

Lrnr_h2o_classifier

Lrnr_h2o_mutator
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.

\code{\link{Lrnr_base}} object with methods for training and
 prediction.

\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides facilities for fitting various types of models with
support for grid search over the hyperparameter space of such models, using
an interface to the H2O platform. For details on the procedures available and
any limitations, consult the documentation of the \code{h2o} package.
}
\keyword{data}
