% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_HarmonicReg.R
\docType{class}
\name{Lrnr_HarmonicReg}
\alias{Lrnr_HarmonicReg}
\title{Harmonic Regression}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_HarmonicReg
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner fits first harmonics in a Fourier expansion to one or more time series.
}
\section{Fields}{

\describe{
\item{\code{n.ahead}}{The forecast horizon. If not specified, returns forecast of size task$X.}

\item{\code{K}}{Maximum order(s) of Fourier terms.}

\item{\code{freq}}{the number of observations per unit of time.}
}}
\seealso{
Other Learners: \code{\link{Lrnr_arima}};
  \code{\link{Lrnr_base}}; \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_glm_fast}},
  \code{\link{Lrnr_h2o_classifier}},
  \code{\link{Lrnr_h2o_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_h2o_mutator}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_pkg_SuperLearner_screener}},
  \code{\link{Lrnr_randomForest}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_xgboost}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_cv}}; \code{\link{Lrnr_delayed}};
  \code{\link{Lrnr_rugarch}}; \code{\link{Lrnr_tsDyn}};
  \code{\link{Pipeline}}; \code{\link{Stack}}
}
\keyword{data}

