% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_condensier_logisfitR6.R
\docType{class}
\name{Lrnr_pkg_condensier_logisfitR6}
\alias{Lrnr_pkg_condensier_logisfitR6}
\title{Lrnr_pkg_condensier_logisfitR6}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_pkg_condensier_logisfitR6
}
\description{
Lrnr_pkg_condensier_logisfitR6
}
\keyword{datasets}
