% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_condensier_logisfitR6.R
\docType{class}
\name{Lrnr_pkg_condensier_logisfitR6}
\alias{Lrnr_pkg_condensier_logisfitR6}
\title{sl3 Learner wrapper for condensier}
\description{
This wrapper allows the use of any \code{sl3} Learner as a Learner for
\code{condensier}. For details, see the \code{\link[condensier]{fit_density}}
function.
}
