% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_pkg_SuperLearner_method.R
\docType{data}
\name{Lrnr_pkg_SuperLearner_method}
\alias{Lrnr_pkg_SuperLearner_method}
\title{Methods for SuperLearner Objects}
\format{\preformatted{Class 'R6ClassGenerator' <Lrnr_pkg_SuperLearner_method> object generator
  Inherits from: <Lrnr_base>
  Public:
    initialize: function (method, ...) 
    clone: function (deep = FALSE) 
  Private:
    .required_packages: SuperLearner
    .train: function (task) 
    .predict: function (task) 
  Parent env: <environment: namespace:sl3>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "Lrnr_pkg_SuperLearner_method_generator"
}}
\usage{
Lrnr_pkg_SuperLearner_method
}
\description{
Support for \code{SuperLearner} wrappers, obviously not the most efficient
approach, we should reimplement as many as possible.
}
\keyword{datasets}

