% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_randomForest.R
\docType{class}
\name{Lrnr_randomForest}
\alias{Lrnr_randomForest}
\title{Random Forests}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_randomForest
}
\value{
Learner object with methods for training and prediction. See \code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides fitting procedures for random forest models, using the
\code{randomForest} package, using the \code{\link[randomForest]{randomForest}} function.
}
\section{Parameters}{

\describe{
\item{\code{ntree=100}}{Number of trees in forest}
\item{\code{keep.forest=TRUE}}{If \code{TRUE}, forest is stored, which is required for prediction.}
\item{\code{nodesize=5}}{Minimum number of observations in terminal (leaf) nodes}
\item{\code{maxnodes=NULL}}{Maximum number of terminal (leaf) nodes in each tree}
\item{\code{importance=FALSE}}{Store variable importance information}
\item{\code{...}}{Other parameters passed to \code{\link[randomForest]{randomForest}}}
}
}

\section{Common Parameters}{


Individual learners have their own sets of parameters. Below is a list of shared parameters, implemented by \code{Lrnr_base}, and shared
by all learners.

\describe{
  \item{\code{covariates}}{A character vector of covariates. The learner will use this to subset the covariates for any specified task}
  \item{\code{outcome_type}}{A \code{\link{variable_type}} object used to control the outcome_type used by the learner. Overrides the task outcome_type if specified}
  \item{\code{...}}{All other parameters should be handled by the invidual learner classes. See the documentation for the learner class you're instantiating}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Lrnr_xgboost}},
  \code{\link{Pipeline}}, \code{\link{Stack}},
  \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
