% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_tsDyn.R
\docType{class}
\name{Lrnr_tsDyn}
\alias{Lrnr_tsDyn}
\title{Nonlinear Time Series Analysis}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_tsDyn
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
This learner supports various forms of nonlinear autoregression, including additive AR, neural nets, SETAR and LSTAR models, threshold VAR and VECM.
}
\section{Fields}{

\describe{
\item{\code{learner}}{Available built-in time series models. Currently available are listed by with availableModels() function.}

\item{\code{m}}{embedding dimension.}

\item{\code{d}}{time delay.}

\item{\code{include}}{Type of deterministic regressors to include.}

\item{\code{type}}{Whether the variable is taken is level, difference or a mix as in the ADF test.}

\item{\code{mL}}{autoregressive order for low regime.}

\item{\code{mH}}{autoregressive order for high regime.}

\item{\code{mM}}{autoregressive order for middle regime.}

\item{\code{thDelay}}{Time delay for the threshold variable.}

\item{\code{common}}{Indicates which elements are common to all regimes.}

\item{\code{ML}}{vector of lags for order for low.}

\item{\code{MM}}{vector of lags for order for middle.}

\item{\code{MH}}{vector of lags for order for high.}

\item{\code{nthresh}}{Threshold of the model.}

\item{\code{trim}}{trimming parameter indicating the minimal percentage of observations in each regime.}

\item{\code{sig}}{significance level for the tests to select the number of regimes.}

\item{\code{control}}{further arguments to be passed as control list to optim.}

\item{\code{r}}{Number of cointegrating relationships.}

\item{\code{model}}{Model to estimate. Choices: VAR/VECM/TAR/MTAR.}

\item{\code{I}}{For VAR only: whether in the VAR variables are to be taken in levels or as a difference.}

\item{\code{beta}}{For VECM only: imposed cointegrating value.}

\item{\code{estim}}{Type of estimator for the VECM (two-step approach or Johansen MLE).}

\item{\code{exogen}}{Inclusion of exogenous variables.}

\item{\code{LRinclude}}{Possibility to include in the long-run relationship and the ECT trend.}

\item{\code{commonInter}}{Whether the deterministic regressors are regime specific.}

\item{\code{mTh}}{combination of variables with same lag order for the transition variable.}

\item{\code{gamma}}{prespecified threshold values.}

\item{\code{dummyToBothRegimes}}{Whether the dummy in the one threshold model is applied to each regime.}

\item{\code{max.iter}}{Number of iterations for the algorithm.}

\item{\code{size}}{number of hidden units in the neural network.}

\item{\code{lag}}{Number of lags to include in each regime.}

\item{\code{n.ahead}}{The forecast horizon.}

\item{\code{ngridBeta}}{number of elements to search for the cointegrating value.}

\item{\code{ngridTh}}{number of elements to search for the threshold value.}

\item{\code{th1}}{different possibilities to pre-specify an exact value, an interval or a central point for the search of the threshold.}

\item{\code{th2}}{different possibilities to pre-specify an exact value or a central point for the search of the second threshold.}

\item{\code{beta0}}{Additional regressors to include in the cointegrating relation.}
}}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_condensier}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_delayed}}, \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\keyword{data}
