% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_xgboost.R
\docType{class}
\name{Lrnr_xgboost}
\alias{Lrnr_xgboost}
\title{xgboost: eXtreme Gradient Boosting}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_xgboost
}
\value{
\code{\link{Lrnr_base}} object with methods for training and
 prediction.
}
\description{
This learner provides fitting procedures for \code{xgboost} models, using the
\code{xgboost} package. Such models are classification and regression trees
with extreme gradient boosting. For details on the fitting procedure, consult
the documentation of the \code{xgboost} package.
}
\keyword{data}

