% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{risk}
\alias{risk}
\title{Risk Esimation}
\usage{
risk(pred, truth, loss = loss_squared_error, weights = NULL)
}
\arguments{
\item{pred}{A vector of predicted values.}

\item{truth}{A vector of true values.}

\item{loss}{A loss function. For options, see \link{loss_functions}.}

\item{weights}{A vector of weights.}
}
\description{
Estimates a risk for a given set of predictions and loss function.
}
