% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared_Data.R
\docType{data}
\name{Shared_Data}
\alias{Shared_Data}
\title{Container Class for data.table Shared Between Tasks}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Shared_Data
}
\description{
Mostly to deal with alloc.col shallow copies, but also nice to have a bit more abstraction.
}
\keyword{datasets}
