% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revere_task.R
\docType{class}
\name{sl3_revere_Task}
\alias{sl3_revere_Task}
\title{Revere (SplitSpecific) Task}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
sl3_revere_Task
}
\description{
A task that has different realizations in different folds
Useful for Revere CV operations
}
\details{
Learners with property "cv" must use these tasks correctly

Other learners will treat this as the equivalent of the "full" task.
}
\keyword{data}
