% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sl3_Task.R
\docType{class}
\name{sl3_Task}
\alias{sl3_Task}
\title{Class for Task Data}
\format{\code{\link{R6Class}} object.}
\usage{
sl3_Task
}
\value{
\code{sl3_Task} object
}
\description{
A thin wrapper around a \code{data.table} containing the data. Contains metadata about the particular machine learning problem, including which variables are to be used as covariates and outcomes.
}
\section{Fields}{

\describe{
\item{\code{data}}{Underlying representation of the data}

\item{\code{nodes}}{A list indicating which columns of \code{data} have which purpose}

\item{\code{X}}{a data.table containing the covariates}

\item{\code{Y}}{a vector containing the outcomes}

\item{\code{weights}}{a vector containing the observation weights. If weights aren't specified on construction, weights will default to 1}

\item{\code{id}}{a vector containing the observation units.}
}}
\section{Methods}{

\describe{
  \item{\code{new(data, covariates, outcome, outcome_type, id, weights, folds, nodes)}}{This method is used to create an object of this class. todo: describe inputs and behavior}
  \item{\code{next_in_chain(new_X)}}{Generates a copy of this task with the set of covariates redefined. This is mostly to be used internally for \code{\link{Pipeline}}s}
  }
}
\keyword{data}

