% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stack.R
\docType{class}
\name{Stack}
\alias{Stack}
\title{Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns. Currently, train fits the learners one-at-a-time, but this will be parallelized with \link{future} going forward.}
\format{\code{\link{R6Class}} object.}
\usage{
Stack
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns. Currently, train fits the learners one-at-a-time, but this will be parallelized with \link{future} going forward.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of learners to stack}
}}
\section{Methods}{

\describe{
  \item{\code{new(...)}}{This method is used to create a stack of learners. Arguments should be indiviual \code{Learner}s.}
  }
}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_arima}}; \code{\link{Lrnr_base}};
  \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_glm_fast}},
  \code{\link{Lrnr_h2o_classifier}},
  \code{\link{Lrnr_h2o_glm}}, \code{\link{Lrnr_h2o_grid}},
  \code{\link{Lrnr_h2o_mutator}}, \code{\link{Lrnr_nnls}},
  \code{\link{Lrnr_pkg_SuperLearner}},
  \code{\link{Lrnr_pkg_SuperLearner_screener}},
  \code{\link{Lrnr_randomForest}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_xgboost}},
  \code{\link{undocumented_learner}};
  \code{\link{Lrnr_cv}}; \code{\link{Lrnr_delayed}};
  \code{\link{Lrnr_rugarch}}; \code{\link{Lrnr_tsDyn}};
  \code{\link{Pipeline}}
}
\keyword{data}

