% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack.R
\docType{class}
\name{Stack}
\alias{Stack}
\title{Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns.}
\format{\code{\link{R6Class}} object.}
\usage{
Stack
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Learner Stacking
A Stack is a special Learner that combines multiple other learners, "stacking" their predictions in columns.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of learners to stack}
}}

\section{Methods}{

\describe{
\item{\code{new(...)}}{This method is used to create a stack of learners. Arguments should be indiviual \code{Learner}s.}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_base}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_sl}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Pipeline}}
}
\keyword{data}
