% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_pkg_SuperLearner.R,
%   R/Lrnr_pkg_SuperLearner_method.R, R/Lrnr_pkg_SuperLearner_screener.R
\docType{class}
\name{Lrnr_pkg_SuperLearner}
\alias{Lrnr_pkg_SuperLearner}
\alias{Lrnr_pkg_SuperLearner_method}
\alias{Lrnr_pkg_SuperLearner_screener}
\title{Use SuperLearner Wrappers, Screeners, and Methods, in sl3}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_pkg_SuperLearner

Lrnr_pkg_SuperLearner_method

Lrnr_pkg_SuperLearner_screener
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
These learners provide an interface to the wrapper functions,
screening algorithms, and combination methods provided by the
\code{SuperLearner} package. These components add support for a range of
algorithms not currently implemented natively in \code{sl3}.

\code{Lrnr_pkg_SuperLearner} - Interface for \code{SuperLearner}
wrapper functions. Use \code{SuperLearner::listWrappers("SL")} for a list.

Use \code{SuperLearner::listWrappers("method")} for a list of options.

Use \code{SuperLearner::listWrappers("screen")} for a list of options.
}
\section{Parameters}{

\describe{
\item{\code{SL_wrapper}}{The wrapper function to use.}
\item{\code{...}}{Currently not used.}
}
}

\section{Common Parameters}{


Individual learners have their own sets of parameters. Below is a list of shared parameters, implemented by \code{Lrnr_base}, and shared
by all learners.

\describe{
  \item{\code{covariates}}{A character vector of covariates. The learner will use this to subset the covariates for any specified task}
  \item{\code{outcome_type}}{A \code{\link{variable_type}} object used to control the outcome_type used by the learner. Overrides the task outcome_type if specified}
  \item{\code{...}}{All other parameters should be handled by the invidual learner classes. See the documentation for the learner class you're instantiating}
}
}

\seealso{
Other Learners: \code{\link{Custom_chain}},
  \code{\link{Lrnr_HarmonicReg}}, \code{\link{Lrnr_arima}},
  \code{\link{Lrnr_bartMachine}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_bilstm}}, \code{\link{Lrnr_condensier}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_dbarts}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_glm_fast}}, \code{\link{Lrnr_glmnet}},
  \code{\link{Lrnr_glm}}, \code{\link{Lrnr_grf}},
  \code{\link{Lrnr_h2o_grid}}, \code{\link{Lrnr_hal9001}},
  \code{\link{Lrnr_independent_binomial}},
  \code{\link{Lrnr_lstm}}, \code{\link{Lrnr_mean}},
  \code{\link{Lrnr_nnls}}, \code{\link{Lrnr_optim}},
  \code{\link{Lrnr_pca}}, \code{\link{Lrnr_randomForest}},
  \code{\link{Lrnr_ranger}}, \code{\link{Lrnr_rpart}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_sl}},
  \code{\link{Lrnr_solnp_density}},
  \code{\link{Lrnr_solnp}},
  \code{\link{Lrnr_subset_covariates}},
  \code{\link{Lrnr_svm}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Lrnr_xgboost}}, \code{\link{Pipeline}},
  \code{\link{Stack}}, \code{\link{define_h2o_X}},
  \code{\link{undocumented_learner}}
}
\keyword{data}
\keyword{datasets}
