% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_h2o_glm.R, R/Lrnr_h2o_grid.R,
%   R/Lrnr_pkg_SuperLearner.R, R/Lrnr_pkg_SuperLearner_method.R,
%   R/Lrnr_pkg_SuperLearner_screener.R, R/Lrnr_sl.R, R/Lrnr_solnp_density.R,
%   R/Lrnr_subset_covariates.R, R/sl3-package.R
\docType{data}
\name{Lrnr_h2o_glm}
\alias{Lrnr_h2o_glm}
\alias{Lrnr_h2o_grid}
\alias{Lrnr_h2o_classifier}
\alias{Lrnr_h2o_mutator}
\alias{Lrnr_pkg_SuperLearner}
\alias{Lrnr_pkg_SuperLearner_method}
\alias{Lrnr_pkg_SuperLearner_screener}
\alias{Lrnr_sl}
\alias{Lrnr_solnp_density}
\alias{Lrnr_subset_covariates}
\alias{undocumented_learner}
\title{Subset Covariate Learner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_h2o_glm

Lrnr_h2o_grid

Lrnr_h2o_classifier

Lrnr_h2o_mutator

Lrnr_pkg_SuperLearner

Lrnr_pkg_SuperLearner_method

Lrnr_pkg_SuperLearner_screener

Lrnr_sl

Lrnr_solnp_density

Lrnr_subset_covariates
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
We haven't documented this one yet.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}
}}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_condensier}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\concept{Learners}
\keyword{data}
\keyword{datasets}
