% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lrnr_sl.R, R/sl3-package.R
\docType{data}
\name{Lrnr_sl}
\alias{Lrnr_sl}
\alias{undocumented_learner}
\title{Undocumented Learner}
\format{\preformatted{Class 'R6ClassGenerator' <Lrnr_sl> object generator
  Inherits from: <Lrnr_base>
  Public:
    initialize: function (learners, metalearner, folds = NULL, keep_extra = TRUE, 
    print: function () 
    metalearner_fit: function () 
    cv_risk: function (loss_fun) 
    clone: function (deep = FALSE) 
  Active bindings:
    name: function () 
  Private:
    .pretrain: function (task) 
    .train: function (task, pretrain) 
    .predict: function (task) 
  Parent env: <environment: namespace:sl3>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "Lrnr_sl_generator"
}}
\usage{
Lrnr_sl
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
We haven't documented this one yet.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}
}}
\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}};
  \code{\link{Lrnr_arima}}; \code{\link{Lrnr_base}};
  \code{\link{Lrnr_condensier}}; \code{\link{Lrnr_cv}};
  \code{\link{Lrnr_define_interactions}};
  \code{\link{Lrnr_expSmooth}}; \code{\link{Lrnr_rugarch}};
  \code{\link{Lrnr_tsDyn}}; \code{\link{Pipeline}};
  \code{\link{Stack}}
}
\keyword{data}
\keyword{datasets}

