% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_sl.R, R/sl3-package.R
\docType{data}
\name{Lrnr_sl}
\alias{Lrnr_sl}
\alias{undocumented_learner}
\title{Undocumented Learner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_sl
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
We haven't documented this one yet.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}
}}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_condensier}}, \code{\link{Lrnr_cv}},
  \code{\link{Lrnr_define_interactions}},
  \code{\link{Lrnr_expSmooth}}, \code{\link{Lrnr_rugarch}},
  \code{\link{Lrnr_tsDyn}}, \code{\link{Pipeline}},
  \code{\link{Stack}}
}
\concept{Learners}
\keyword{data}
\keyword{datasets}
