% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_condensier.R,
%   R/Lrnr_define_interactions.R, R/Lrnr_glm.R, R/Lrnr_glm_fast.R,
%   R/Lrnr_h2o_glm.R, R/Lrnr_h2o_grid.R, R/Lrnr_nnls.R,
%   R/Lrnr_pkg_SuperLearner.R, R/Lrnr_pkg_SuperLearner_screener.R,
%   R/Lrnr_randomForest.R, R/Lrnr_sl.R, R/Lrnr_solnp_density.R,
%   R/Lrnr_subset_covariates.R, R/Lrnr_xgboost.R, R/sl3-package.R
\docType{data}
\name{Lrnr_condensier}
\alias{Lrnr_condensier}
\alias{Lrnr_define_interactions}
\alias{Lrnr_glm}
\alias{Lrnr_glm_fast}
\alias{Lrnr_h2o_glm}
\alias{Lrnr_h2o_grid}
\alias{Lrnr_h2o_classifier}
\alias{Lrnr_h2o_mutator}
\alias{Lrnr_nnls}
\alias{Lrnr_pkg_SuperLearner}
\alias{Lrnr_pkg_SuperLearner_screener}
\alias{Lrnr_randomForest}
\alias{Lrnr_sl}
\alias{Lrnr_solnp_density}
\alias{Lrnr_subset_covariates}
\alias{Lrnr_xgboost}
\alias{undocumented_learner}
\title{Subset Covariate Learner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_condensier

Lrnr_define_interactions

Lrnr_glm

Lrnr_glm_fast

Lrnr_h2o_glm

Lrnr_h2o_grid

Lrnr_h2o_classifier

Lrnr_h2o_mutator

Lrnr_nnls

Lrnr_pkg_SuperLearner

Lrnr_pkg_SuperLearner_screener

Lrnr_randomForest

Lrnr_sl

Lrnr_solnp_density

Lrnr_subset_covariates

Lrnr_xgboost
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Subset Covariate Learner

We haven't documented this one yet.
}
\section{Fields}{

\describe{
\item{\code{params}}{A list of parameters needed to fully specify the learner. This includes things like model hyperparameters.}
}}

\seealso{
Other Learners: \code{\link{Lrnr_HarmonicReg}},
  \code{\link{Lrnr_arima}}, \code{\link{Lrnr_base}},
  \code{\link{Lrnr_cv}}, \code{\link{Lrnr_expSmooth}},
  \code{\link{Lrnr_rugarch}}, \code{\link{Lrnr_tsDyn}},
  \code{\link{Pipeline}}, \code{\link{Stack}}
}
\keyword{data}
\keyword{datasets}
