% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable_Type.R
\docType{data}
\name{Variable_Type}
\alias{Variable_Type}
\alias{variable_type}
\title{Specify variable type}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Variable_Type

variable_type(type = NULL, levels = NULL, bounds = NULL, x = NULL,
  pcontinuous = 0.05)
}
\arguments{
\item{type}{A type name}

\item{levels}{valid levels for discrete types}

\item{bounds}{bounds for continous variables}

\item{x}{data to use for inferring type if not specified}

\item{pcontinuous}{if type inferred, proportion of unique observations above which variable is continuous}

\item{...}{not currently used}
}
\description{
Specify variable type
}
\keyword{datasets}
