% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable_Type.R
\name{Variable_Type}
\alias{Variable_Type}
\alias{variable_type}
\title{Specify variable type}
\usage{
variable_type(type = NULL, levels = NULL, bounds = NULL, x = NULL,
  pcontinuous = getOption("sl3.pcontinuous"))
}
\arguments{
\item{type}{A type name.}

\item{levels}{Valid levels for discrete types.}

\item{bounds}{Bounds for continous variables.}

\item{x}{Data to use for inferring type if not specified.}

\item{pcontinuous}{If \code{type} above is inferred, the proportion of unique
observations above which variable is continuous}
}
\description{
Specify variable type
}
