% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Variable Importance Measure}
\usage{
varimp(fit, loss, type = c("permute", "sample"),
  fold_number = "validation")
}
\arguments{
\item{fit}{A learner fitted task}

\item{loss}{A loss function (see loss_functions.R)}

\item{type}{either "permute" (for permutation based variable importance),
or "sample" (for sample based variable importance)}

\item{fold_number}{either "full" for full fit vim, "validation" for
cross-validated vim, or a positive integer for fold-specific vim}
}
\value{
A table of risk differences for each covariate, where a higher risk
difference indicates a more important covariate in predicting Y
}
\description{
Function that takes a learner fitted task and loss function to generate a
table where each row corresponds to the risk when we omit a covariate minus
the risk when we do not omit the covariate. Specifically, we calculate the
risk difference between the learner fit with a scrambled covariate and the
leaner fit with the true covariate, across all covariates.
}
\keyword{importance}
\keyword{variable}
