% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_learner_template}
\alias{write_learner_template}
\title{Generate a file containing a template \code{sl3} Learner}
\usage{
write_learner_template(file)
}
\arguments{
\item{file}{the path where the file should be written}
}
\value{
the return from \code{\link{file.copy}}. \code{TRUE} if writing the
template was successful.
}
\description{
Generates a template file that can be used to write a new \code{sl3} Learner.
For more information, see the \href{../doc/custom_lrnrs.html}{Defining New
Learners} vignette.
}
